refactor(cloudsave): 拆 utils/cloudsave_runtime.py 上帝文件为领域子模块包
重构路线图第一批第 1 刀。4224 行/176 个顶层符号按 #2148 已验证的
「包 + _shared.py + __init__ 门面 re-export」模式拆为 8 个子模块:
_shared(常量/异常/闸门/审核/敏感扫描)、fence(root mode+写栅栏+
跨进程锁,锁全局唯一)、staging(staging/哈希/sqlite shadow/tombstone)、
legacy_migration、bindings、snapshots、operations、bootstrap。

纯搬移不改逻辑,AST 级符号对账:旧文件全部顶层符号在包内逐符号
AST 等价,唯 7 个函数改为经门面晚绑定调用 _atomic_copy_file /
_apply_runtime_file / set_root_mode,保住测试对包门面的 monkeypatch
语义。门面 re-export 全部旧符号,main/test 引用零改写(仅
steam_cloud_bundle 测试 4 处 patch 目标跟随 helper 转正改公开名)。

steam_cloud_bundle 在用的四个私有 helper 转正公开名:
assert_deadline_not_exceeded / atomic_copy_file /
create_staging_workspace / load_json_if_exists(私有名保留)。

验证:pytest tests/unit 失败集合与 origin/main 基线一致
(61 failed 存量红/4995 passed);check_startup_import_lazy、
check_docstring_no_cjk、ruff 全过。
