--- utils/cloudsave_runtime.py
@@ -1,4224 +1,0 @@
-# Copyright 2025-2026 Project N.E.K.O. Team
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from __future__ import annotations
-
-import hashlib
-import json
-import logging
-import os
-import re
-import shutil
-import sqlite3
-import sys
-import tempfile
-import time
-import unicodedata
-from copy import deepcopy
-from contextlib import contextmanager, nullcontext
-from datetime import datetime, timezone
-from pathlib import Path
-from typing import Any
-
-from config import CHARACTER_RESERVED_FIELDS, DEFAULT_CONFIG_DATA
-from utils.character_name import PROFILE_NAME_MAX_UNITS, validate_character_name
-from utils.file_utils import atomic_write_json
-from utils.storage_path_rewrite import rebase_runtime_bound_workshop_config_paths
-
-
-logger = logging.getLogger(__name__)
-
-ROOT_MODE_NORMAL = "normal"
-ROOT_MODE_BOOTSTRAP_IMPORTING = "bootstrap_importing"
-ROOT_MODE_BOOTSTRAP_READONLY = "bootstrap_readonly"
-ROOT_MODE_DEFERRED_INIT = "deferred_init"
-ROOT_MODE_MAINTENANCE_READONLY = "maintenance_readonly"
-CLOUDSAVE_DISABLED_ENV = "NEKO_CLOUDSAVE_DISABLED"
-CLOUDSAVE_DISABLED_LOCAL_STATE_UNAVAILABLE = "local_state_unavailable"
-
-WRITE_BLOCKING_MODES = frozenset(
-    {
-        ROOT_MODE_BOOTSTRAP_IMPORTING,
-        ROOT_MODE_BOOTSTRAP_READONLY,
-        ROOT_MODE_DEFERRED_INIT,
-        ROOT_MODE_MAINTENANCE_READONLY,
-    }
-)
-
-SENSITIVE_TOKENS = (
-    "api_key",
-    "authorization",
-    "bearer",
-    "cookie",
-    "token",
-    "sk-",
-)
-SENSITIVE_KEY_NAMES = frozenset(
-    {
-        "api_key",
-        "apikey",
-        "authorization",
-        "cookie",
-        "cookies",
-        "token",
-        "access_token",
-        "refresh_token",
-        "id_token",
-        "session_token",
-        "auth_token",
-        "bearer_token",
-        "sessionid",
-        "session_id",
-    }
-)
-SENSITIVE_VALUE_PATTERNS = (
-    re.compile(r"\bsk-[A-Za-z0-9][A-Za-z0-9._-]{12,}\b"),
-    re.compile(r"\bbearer\s+[A-Za-z0-9._-]{12,}\b", re.IGNORECASE),
-    re.compile(r"\b(?:api[_\-\s]*key|authorization|cookie|token)\s*[:=]\s*[^\s]{8,}\b", re.IGNORECASE),
-)
-
-GLOBAL_CONVERSATION_KEY = "__global_conversation__"
-MANAGED_MEMORY_FILENAMES = (
-    "recent.json",
-    "settings.json",
-    "facts.json",
-    "facts_archive.json",
-    "persona.json",
-    "persona_corrections.json",
-    "reflections.json",
-    "reflections_archive.json",
-    "surfaced.json",
-    "time_indexed.db",
-)
-MANAGED_CLOUDSAVE_PREFIXES = (
-    "characters/",
-    "catalog/",
-    "profiles/",
-    "bindings/",
-    "memory/",
-    "overrides/",
-    "meta/",
-)
-LEGACY_RUNTIME_DIR_NAMES = (
-    "config",
-    "memory",
-    "plugins",
-    "live2d",
-    "vrm",
-    "mmd",
-    "workshop",
-    "character_cards",
-    "card_faces",
-    "cloudsave",
-    "cloudsave_backups",
-    ".cloudsave_staging",
-)
-NON_RUNTIME_CONTENT_DIR_NAMES = {
-    "cloudsave",
-    "cloudsave_backups",
-    ".cloudsave_staging",
-}
-LEGACY_OPTIONAL_STATE_FILES = (
-    "cloudsave_local_state.json",
-)
-TARGET_OPTIONAL_STATE_FILES = (
-    "root_state.json",
-    "cloudsave_local_state.json",
-    "character_tombstones.json",
-)
-ROOT_CONFIG_MERGE_FILES = (
-    "core_config.json",
-    "voice_storage.json",
-    "workshop_config.json",
-)
-RUNTIME_ASSET_DIR_NAMES = (
-    "plugins",
-    "live2d",
-    "vrm",
-    "mmd",
-    "workshop",
-    "character_cards",
-    "card_faces",
-)
-
-_cloud_apply_lock_handle = None
-_cloud_apply_lock_file = None
-SQLITE_FILE_HEADER = b"SQLite format 3\x00"
-
-
-class MaintenanceModeError(RuntimeError):
-    """Raised when a write is attempted while the global cloudsave fence is active."""
-
-    def __init__(self, mode: str, *, operation: str = "write", target: str = ""):
-        self.mode = str(mode or ROOT_MODE_NORMAL)
-        self.operation = str(operation or "write")
-        self.target = str(target or "")
-        self.code = "CLOUDSAVE_WRITE_FENCE_ACTIVE"
-        detail = f"{self.operation} blocked while root_state.mode={self.mode}"
-        if self.target:
-            detail = f"{detail} ({self.target})"
-        super().__init__(detail)
-
-
-class CloudsaveOperationError(RuntimeError):
-    """Raised when a single-character cloudsave operation cannot proceed safely."""
-
-    def __init__(self, code: str, message: str, *, character_name: str = ""):
-        self.code = str(code or "CLOUDSAVE_OPERATION_FAILED")
-        self.character_name = str(character_name or "")
-        super().__init__(message)
-
-
-class CloudsaveDeadlineExceeded(RuntimeError):
-    """Raised when a cloudsave job exceeds its pre-apply time budget."""
-
-    def __init__(self, operation: str, stage: str):
-        self.operation = str(operation or "cloudsave")
-        self.stage = str(stage or "unknown")
-        self.code = "CLOUDSAVE_DEADLINE_EXCEEDED"
-        super().__init__(f"{self.operation} exceeded deadline before stage={self.stage}")
-
-
-def _utc_now_iso() -> str:
-    return datetime.now(timezone.utc).isoformat().replace("+00:00", "Z")
-
-
-def _assert_deadline_not_exceeded(
-    deadline_monotonic: float | None,
-    *,
-    operation: str,
-    stage: str,
-) -> None:
-    if deadline_monotonic is None:
-        return
-    if time.monotonic() <= float(deadline_monotonic):
-        return
-    raise CloudsaveDeadlineExceeded(operation=operation, stage=stage)
-
-
-def is_cloudsave_provider_available(config_manager) -> bool:
-    """Centralize provider availability so future remote probes only need one hook."""
-    if is_cloudsave_disabled():
-        return False
-    override = getattr(config_manager, "cloudsave_provider_available", None)
-    if override is None:
-        return True
-    return bool(override)
-
-
-def cloudsave_disabled_reason() -> str:
-    return str(os.environ.get(CLOUDSAVE_DISABLED_ENV) or "").strip()
-
-
-def is_cloudsave_disabled() -> bool:
-    return bool(cloudsave_disabled_reason())
-
-
-def is_cloudsave_disabled_due_to_local_state_unavailable() -> bool:
-    return cloudsave_disabled_reason() == CLOUDSAVE_DISABLED_LOCAL_STATE_UNAVAILABLE
-
-
-def _raise_cloudsave_disabled(operation: str, *, character_name: str = "") -> None:
-    reason = cloudsave_disabled_reason() or "unknown"
-    raise CloudsaveOperationError(
-        "CLOUDSAVE_PROVIDER_UNAVAILABLE",
-        f"Cloudsave is disabled for this session ({reason}); skipped {operation}.",
-        character_name=character_name,
-    )
-
-
-def build_default_cloudsave_manifest(*, client_id: str = "") -> dict[str, Any]:
-    """Build the minimal local manifest skeleton for phase 0."""
-    return {
-        "schema_version": 1,
-        "min_reader_schema_version": 1,
-        "min_app_version": "",
-        "client_id": str(client_id or ""),
-        "device_id": "",
-        "sequence_number": 0,
-        "exported_at_utc": "",
-        "files": {},
-        "fingerprint": "",
-    }
-
-
-def _json_canonical_dumps(payload: Any) -> str:
-    return json.dumps(payload, ensure_ascii=False, sort_keys=True, separators=(",", ":"))
-
-
-def _sha256_bytes(payload: bytes) -> str:
-    return hashlib.sha256(payload).hexdigest()
-
-
-def _sha256_file(path: Path) -> str:
-    digest = hashlib.sha256()
-    with open(path, "rb") as file_obj:
-        for chunk in iter(lambda: file_obj.read(1024 * 1024), b""):
-            digest.update(chunk)
-    return digest.hexdigest()
-
-
-def _create_staging_workspace(config_manager, prefix: str) -> Path:
-    config_manager.ensure_cloudsave_structure()
-    return Path(
-        tempfile.mkdtemp(
-            prefix=f"{prefix}-",
-            dir=str(config_manager.cloudsave_staging_dir),
-        )
-    )
-
-
-def _atomic_copy_file(source_path: Path, target_path: Path) -> None:
-    target_path.parent.mkdir(parents=True, exist_ok=True)
-    fd, temp_path = tempfile.mkstemp(
-        prefix=f".{target_path.name}.",
-        suffix=".tmp",
-        dir=str(target_path.parent),
-    )
-    try:
-        with os.fdopen(fd, "wb") as temp_file, open(source_path, "rb") as source_file:
-            shutil.copyfileobj(source_file, temp_file)
-            temp_file.flush()
-            os.fsync(temp_file.fileno())
-        os.replace(temp_path, target_path)
-    except Exception:
-        try:
-            os.remove(temp_path)
-        except FileNotFoundError:
-            pass
-        raise
-
-
-def _stage_json_file(stage_root: Path, relative_path: str, payload: Any) -> Path:
-    target_path = stage_root / relative_path
-    atomic_write_json(target_path, payload, ensure_ascii=False, indent=2)
-    return target_path
-
-
-def _stage_file_copy(stage_root: Path, relative_path: str, source_path: Path) -> Path:
-    staged_path = stage_root / relative_path
-    staged_path.parent.mkdir(parents=True, exist_ok=True)
-    shutil.copy2(source_path, staged_path)
-    return staged_path
-
-
-def _looks_like_sqlite_database(source_path: Path) -> bool:
-    try:
-        if source_path.stat().st_size < len(SQLITE_FILE_HEADER):
-            return False
-        with open(source_path, "rb") as file_obj:
-            return file_obj.read(len(SQLITE_FILE_HEADER)) == SQLITE_FILE_HEADER
-    except OSError:
-        return False
-
-
-def _run_sqlite_shadow_copy(source_path: Path, target_path: Path) -> None:
-    target_path.parent.mkdir(parents=True, exist_ok=True)
-    journal_mode = ""
-
-    with sqlite3.connect(str(source_path), timeout=5.0, isolation_level=None) as source_conn:
-        source_conn.execute("PRAGMA busy_timeout = 5000")
-        try:
-            row = source_conn.execute("PRAGMA journal_mode").fetchone()
-            journal_mode = str(row[0]).lower() if row and row[0] is not None else ""
-        except sqlite3.DatabaseError:
-            journal_mode = ""
-
-        if journal_mode == "wal":
-            try:
-                checkpoint_row = source_conn.execute("PRAGMA wal_checkpoint(TRUNCATE)").fetchone() or ()
-                if checkpoint_row and int(checkpoint_row[0] or 0) != 0:
-                    logger.warning(
-                        "SQLite wal_checkpoint(TRUNCATE) reported busy=%s for %s; continuing with backup API",
-                        checkpoint_row[0],
-                        source_path,
-                    )
-            except sqlite3.DatabaseError as exc:
-                logger.warning(
-                    "SQLite wal_checkpoint(TRUNCATE) failed for %s; continuing with backup API: %s",
-                    source_path,
-                    exc,
-                )
-
-        with sqlite3.connect(str(target_path), timeout=5.0, isolation_level=None) as target_conn:
-            target_conn.execute("PRAGMA busy_timeout = 5000")
-            source_conn.backup(target_conn)
-            quick_check = target_conn.execute("PRAGMA quick_check").fetchone()
-            quick_check_result = str(quick_check[0]) if quick_check and quick_check[0] is not None else ""
-            if quick_check_result.lower() != "ok":
-                raise sqlite3.DatabaseError(
-                    f"shadow copy integrity check failed for {source_path}: {quick_check_result or 'unknown'}"
-                )
-
-
-def _stage_memory_file(stage_root: Path, relative_path: str, source_path: Path) -> Path:
-    if source_path.name != "time_indexed.db" or not _looks_like_sqlite_database(source_path):
-        return _stage_file_copy(stage_root, relative_path, source_path)
-
-    staged_path = stage_root / relative_path
-    try:
-        _run_sqlite_shadow_copy(source_path, staged_path)
-        return staged_path
-    except sqlite3.DatabaseError as exc:
-        raise RuntimeError(f"failed to create SQLite shadow copy for {source_path}: {exc}") from exc
-
-
-def _apply_runtime_file(source_path: Path, target_path: Path) -> None:
-    if source_path.name == "time_indexed.db" and _looks_like_sqlite_database(source_path):
-        target_looks_like_sqlite = not target_path.exists() or _looks_like_sqlite_database(target_path)
-        if target_looks_like_sqlite:
-            try:
-                _run_sqlite_shadow_copy(source_path, target_path)
-                return
-            except sqlite3.DatabaseError as exc:
-                raise RuntimeError(f"failed to apply SQLite backup copy for {target_path}: {exc}") from exc
-
-    _atomic_copy_file(source_path, target_path)
-
-
-def _list_existing_cloudsave_files(config_manager) -> set[str]:
-    existing_files: set[str] = set()
-    for prefix in MANAGED_CLOUDSAVE_PREFIXES:
-        prefix_path = config_manager.cloudsave_dir / prefix.rstrip("/")
-        if not prefix_path.exists():
-            continue
-        for file_path in prefix_path.rglob("*"):
-            if file_path.is_file():
-                existing_files.add(str(file_path.relative_to(config_manager.cloudsave_dir)).replace("\\", "/"))
-    return existing_files
-
-
-def _cleanup_empty_parent_dirs(path: Path, stop_at: Path) -> None:
-    current = path.parent
-    while current != stop_at and current.exists():
-        try:
-            current.rmdir()
-        except OSError:
-            break
-        current = current.parent
-
-
-def _build_manifest_fingerprint(*, client_id: str, sequence_number: int, files: dict[str, Any]) -> str:
-    payload = {
-        "client_id": client_id,
-        "sequence_number": int(sequence_number),
-        "files": files,
-    }
-    return _sha256_bytes(_json_canonical_dumps(payload).encode("utf-8"))
-
-
-def _normalize_tombstone_entry(entry: Any) -> dict[str, Any] | None:
-    if not isinstance(entry, dict):
-        return None
-    character_name = str(entry.get("character_name") or entry.get("name") or "").strip()
-    if not character_name:
-        return None
-    try:
-        sequence_number = int(entry.get("sequence_number") or 0)
-    except (TypeError, ValueError):
-        sequence_number = 0
-    return {
-        "character_name": character_name,
-        "deleted_at": str(entry.get("deleted_at") or ""),
-        "sequence_number": sequence_number,
-    }
-
-
-def _normalize_tombstones_state(payload: Any) -> dict[str, Any]:
-    raw_entries = []
-    if isinstance(payload, dict):
-        raw_entries = payload.get("tombstones") or []
-    elif isinstance(payload, list):
-        raw_entries = payload
-
-    normalized_entries: dict[str, dict[str, Any]] = {}
-    for raw_entry in raw_entries:
-        normalized_entry = _normalize_tombstone_entry(raw_entry)
-        if normalized_entry is None:
-            continue
-        key = normalized_entry["character_name"]
-        existing_entry = normalized_entries.get(key)
-        if existing_entry is None or normalized_entry["sequence_number"] >= existing_entry["sequence_number"]:
-            normalized_entries[key] = normalized_entry
-
-    return {
-        "version": 1,
-        "tombstones": [
-            normalized_entries[name]
-            for name in sorted(normalized_entries)
-        ],
-    }
-
-
-def _load_local_tombstones_state(config_manager) -> dict[str, Any]:
-    return _normalize_tombstones_state(config_manager.load_character_tombstones_state())
-
-
-def _save_local_tombstones_state(config_manager, payload: Any) -> dict[str, Any]:
-    normalized_state = _normalize_tombstones_state(payload)
-    config_manager.save_character_tombstones_state(normalized_state)
-    return normalized_state
-
-
-def _load_tombstone_names_from_state_path(state_path: Path) -> set[str]:
-    payload = _load_json_if_exists(state_path)
-    normalized_state = _normalize_tombstones_state(payload)
-    return {
-        entry["character_name"]
-        for entry in normalized_state.get("tombstones") or []
-        if isinstance(entry, dict) and entry.get("character_name")
-    }
-
-
-def _make_tombstones_catalog_payload(*, tombstones: list[dict[str, Any]], sequence_number: int, exported_at: str) -> dict[str, Any]:
-    return {
-        "schema_version": 1,
-        "sequence_number": sequence_number,
-        "exported_at_utc": exported_at,
-        "tombstones": deepcopy(tombstones),
-    }
-
-
-def _normalize_audit_name(raw_name: Any) -> str:
-    return unicodedata.normalize("NFC", str(raw_name or "").strip())
-
-
-def audit_cloudsave_character_names(
-    character_names: list[str] | tuple[str, ...],
-    tombstone_names: list[str] | tuple[str, ...] = (),
-) -> dict[str, Any]:
-    errors: list[dict[str, Any]] = []
-    warnings: list[dict[str, Any]] = []
-    entries_by_key: dict[str, list[dict[str, Any]]] = {}
-
-    def _record_entry(source: str, raw_name: Any):
-        original = "" if raw_name is None else str(raw_name)
-        trimmed = original.strip()
-        normalized = _normalize_audit_name(original)
-
-        if original != trimmed:
-            errors.append({
-                "type": "trimmed_whitespace",
-                "source": source,
-                "name": original,
-            })
-
-        validation = validate_character_name(
-            trimmed,
-            # Cloudsave paths legitimately use names like "N.E.K.O" in both
-            # directory names and legacy "*.json" mirrors. Keep the broader
-            # filesystem safety checks, but allow embedded dots here.
-            allow_dots=True,
-            max_units=PROFILE_NAME_MAX_UNITS,
-        )
-        if not validation.ok:
-            errors.append({
-                "type": "invalid_name",
-                "source": source,
-                "name": original,
-                "code": validation.code,
-                "invalid_char": validation.invalid_char,
-            })
-
-        if trimmed and normalized != trimmed:
-            warnings.append({
-                "type": "normalization_changed",
-                "source": source,
-                "name": original,
-                "normalized_name": normalized,
-            })
-
-        if normalized:
-            casefold_key = normalized.casefold()
-            entries_by_key.setdefault(casefold_key, []).append({
-                "source": source,
-                "name": original,
-                "normalized_name": normalized,
-            })
-
-    for name in character_names:
-        _record_entry("character", name)
-    for name in tombstone_names:
-        _record_entry("tombstone", name)
-
-    for casefold_key, entries in entries_by_key.items():
-        normalized_names = {entry["normalized_name"] for entry in entries}
-        original_names = {entry["name"] for entry in entries}
-        if len(entries) > 1 and (len(normalized_names) > 1 or len(original_names) > 1):
-            errors.append({
-                "type": "casefold_conflict",
-                "casefold_key": casefold_key,
-                "entries": entries,
-            })
-
-    return {
-        "ok": not errors,
-        "errors": errors,
-        "warnings": warnings,
-    }
-
-
-def _raise_for_name_audit(audit_result: dict[str, Any], *, context: str) -> None:
-    errors = audit_result.get("errors") or []
-    if not errors:
-        return
-
-    rendered_errors = []
-    for error in errors[:5]:
-        error_type = error.get("type")
-        if error_type == "casefold_conflict":
-            rendered_errors.append(
-                "casefold_conflict:"
-                + ",".join(f"{entry.get('source')}={entry.get('name')}" for entry in error.get("entries") or [])
-            )
-        elif error_type == "invalid_name":
-            rendered_errors.append(
-                f"invalid_name:{error.get('source')}={error.get('name')}({error.get('code')})"
-            )
-        else:
-            rendered_errors.append(f"{error_type}:{error.get('source')}={error.get('name')}")
-    raise ValueError(f"{context} character name audit failed: {'; '.join(rendered_errors)}")
-
-
-def _runtime_config_path_matches_pristine_default(config_manager, runtime_path: Path) -> bool:
-    source_path = None
-    if runtime_path.name == "characters.json":
-        localized_source = getattr(config_manager, "_get_localized_characters_source", lambda: None)()
-        if localized_source:
-            source_path = Path(localized_source)
-    if source_path is None:
-        candidate = Path(config_manager.project_config_dir) / runtime_path.name
-        if candidate.exists():
-            source_path = candidate
-
-    if source_path is not None and source_path.exists():
-        try:
-            return runtime_path.read_bytes() == source_path.read_bytes()
-        except Exception:
-            return False
-
-    default_payload = DEFAULT_CONFIG_DATA.get(runtime_path.name)
-    if default_payload is None:
-        return False
-    try:
-        return json.loads(runtime_path.read_text(encoding="utf-8")) == default_payload
-    except Exception:
-        return False
-
-
-def _runtime_config_dir_has_user_content(config_manager) -> bool:
-    config_dir = Path(config_manager.config_dir)
-    if not config_dir.exists():
-        return False
-    for child in config_dir.iterdir():
-        if _is_ignorable_runtime_entry(child):
-            continue
-        if child.is_dir():
-            return True
-        if not _runtime_config_path_matches_pristine_default(config_manager, child):
-            return True
-    return False
-
-
-def _runtime_root_has_user_content(root: Path, *, config_manager=None) -> bool:
-    if not root.exists():
-        return False
-    config_dir = None
-    if config_manager is not None:
-        try:
-            config_dir = Path(config_manager.config_dir)
-        except Exception:
-            config_dir = None
-    for name in LEGACY_RUNTIME_DIR_NAMES:
-        if name in NON_RUNTIME_CONTENT_DIR_NAMES:
-            continue
-        candidate = root / name
-        if candidate.is_file():
-            return True
-        if candidate.is_dir():
-            if config_dir is not None and candidate == config_dir:
-                if _runtime_config_dir_has_user_content(config_manager):
-                    return True
-                continue
-            try:
-                for child in candidate.iterdir():
-                    if _is_ignorable_runtime_entry(child):
-                        continue
-                    return True
-            except StopIteration:
-                continue
-    return False
-
-
-def runtime_root_has_user_content(root: Path, *, config_manager=None) -> bool:
-    """Public wrapper for detecting user-owned runtime data in a storage root."""
-    return _runtime_root_has_user_content(root, config_manager=config_manager)
-
-
-def _is_ignorable_runtime_entry(path: Path) -> bool:
-    name = path.name
-    if name == ".gitkeep":
-        return True
-    if name.startswith("."):
-        return True
-    if name == "__pycache__":
-        return True
-    return False
-
-
-def _copy_runtime_root_entries(source_root: Path, destination_root: Path) -> list[str]:
-    copied_paths: list[str] = []
-    for name in LEGACY_RUNTIME_DIR_NAMES:
-        source_path = source_root / name
-        if not source_path.exists():
-            continue
-        destination_path = destination_root / name
-        if source_path.is_dir():
-            shutil.copytree(source_path, destination_path, dirs_exist_ok=True)
-        else:
-            destination_path.parent.mkdir(parents=True, exist_ok=True)
-            shutil.copy2(source_path, destination_path)
-        copied_paths.append(name)
-    return copied_paths
-
-
-def _load_json_if_exists(path: Path) -> Any | None:
-    if not path.is_file():
-        return None
-    try:
-        with open(path, "r", encoding="utf-8") as file_obj:
-            return json.load(file_obj)
-    except Exception:
-        return None
-
-
-def _directory_has_meaningful_content(path: Path) -> bool:
-    if not path.is_dir():
-        return False
-    try:
-        for child in path.iterdir():
-            if _is_ignorable_runtime_entry(child):
-                continue
-            return True
-    except Exception:
-        return False
-    return False
-
-
-def _collect_memory_character_names(root: Path) -> set[str]:
-    memory_root = root / "memory"
-    character_names: set[str] = set()
-    if not memory_root.is_dir():
-        return character_names
-    try:
-        for child in memory_root.iterdir():
-            if _is_ignorable_runtime_entry(child):
-                continue
-            if child.is_dir() and _directory_has_meaningful_content(child):
-                character_names.add(child.name)
-            elif child.is_file():
-                character_names.add(child.stem)
-    except Exception:
-        return character_names
-    return character_names
-
-
-def _load_seed_characters_payload(config_manager) -> dict[str, Any]:
-    localized_source = None
-    try:
-        localized_source = config_manager._get_localized_characters_source()
-    except Exception:
-        localized_source = None
-    if localized_source is not None:
-        payload = _load_json_if_exists(Path(localized_source))
-        if isinstance(payload, dict):
-            return payload
-    fallback_payload = config_manager.get_default_characters()
-    return fallback_payload if isinstance(fallback_payload, dict) else {}
-
-
-def _normalize_catgirl_payload(payload: Any) -> dict[str, Any] | None:
-    if not isinstance(payload, dict):
-        return None
-    normalized_payload = deepcopy(payload)
-    try:
-        from utils.config_manager import migrate_catgirl_reserved
-
-        migrate_catgirl_reserved(normalized_payload)
-    except Exception:
-        pass
-    return normalized_payload
-
-
-def _character_payload_looks_default(config_manager, name: str, payload: Any) -> bool:
-    normalized_payload = _normalize_catgirl_payload(payload)
-    if normalized_payload is None:
-        return False
-    default_payload = _normalize_catgirl_payload((_load_seed_characters_payload(config_manager).get("猫娘") or {}).get(name))
-    return default_payload is not None and normalized_payload == default_payload
-
-
-def _master_payload_looks_default(config_manager, payload: Any) -> bool:
-    if not isinstance(payload, dict):
-        return False
-    default_payload = _load_seed_characters_payload(config_manager).get("主人")
-    return default_payload is not None and payload == default_payload
-
-
-def _normalize_preferences_payload(payload: Any) -> list[Any]:
-    if isinstance(payload, list):
-        return deepcopy(payload)
-    if isinstance(payload, dict):
-        return [deepcopy(payload)]
-    return []
-
-
-def _preferences_entry_key(entry: Any) -> str:
-    if isinstance(entry, dict) and entry.get("model_path") is not None:
-        return f"model_path:{entry.get('model_path')}"
-    return _json_canonical_dumps(entry)
-
-
-def _merge_preferences_payloads(legacy_payload: Any, current_payload: Any) -> list[Any]:
-    merged_entries: dict[str, Any] = {}
-    ordered_keys: list[str] = []
-    for payload in (_normalize_preferences_payload(legacy_payload), _normalize_preferences_payload(current_payload)):
-        for entry in payload:
-            key = _preferences_entry_key(entry)
-            if key not in merged_entries:
-                ordered_keys.append(key)
-            merged_entries[key] = deepcopy(entry)
-    return [merged_entries[key] for key in ordered_keys]
-
-
-def _deep_merge_json_dicts(legacy_payload: Any, current_payload: Any) -> dict[str, Any]:
-    legacy_dict = deepcopy(legacy_payload) if isinstance(legacy_payload, dict) else {}
-    current_dict = current_payload if isinstance(current_payload, dict) else {}
-    for key, value in current_dict.items():
-        if isinstance(legacy_dict.get(key), dict) and isinstance(value, dict):
-            legacy_dict[key] = _deep_merge_json_dicts(legacy_dict[key], value)
-        else:
-            legacy_dict[key] = deepcopy(value)
-    return legacy_dict
-
-
-def _config_payload_looks_default(filename: str, payload: Any) -> bool:
-    default_payload = DEFAULT_CONFIG_DATA.get(filename)
-    if filename == "user_preferences.json":
-        return _normalize_preferences_payload(payload) == _normalize_preferences_payload(default_payload)
-    if isinstance(default_payload, dict):
-        return isinstance(payload, dict) and deepcopy(payload) == deepcopy(default_payload)
-    if isinstance(default_payload, list):
-        return isinstance(payload, list) and deepcopy(payload) == deepcopy(default_payload)
-    return False
-
-
-def _config_payload_looks_seeded(config_manager, filename: str, payload: Any) -> bool:
-    project_payload = _load_json_if_exists(Path(config_manager.project_config_dir) / filename)
-    if project_payload is not None:
-        if filename == "user_preferences.json":
-            return _normalize_preferences_payload(payload) == _normalize_preferences_payload(project_payload)
-        return deepcopy(payload) == deepcopy(project_payload)
-    return _config_payload_looks_default(filename, payload)
-
-
-def _merge_characters_payloads(
-    config_manager,
-    legacy_payload: Any,
-    current_payload: Any,
-    *,
-    preserve_current_only_defaults: bool,
-) -> dict[str, Any]:
-    legacy_dict = deepcopy(legacy_payload) if isinstance(legacy_payload, dict) else {}
-    current_dict = deepcopy(current_payload) if isinstance(current_payload, dict) else {}
-    merged_payload = deepcopy(legacy_dict)
-
-    for key, value in current_dict.items():
-        if key not in {"猫娘", "主人", "当前猫娘"}:
-            merged_payload[key] = deepcopy(value)
-
-    legacy_catgirls = legacy_dict.get("猫娘") or {}
-    current_catgirls = current_dict.get("猫娘") or {}
-    merged_catgirls: dict[str, Any] = {}
-    for name in sorted(set(legacy_catgirls) | set(current_catgirls)):
-        legacy_character = legacy_catgirls.get(name)
-        current_character = current_catgirls.get(name)
-        if legacy_character is None:
-            if not preserve_current_only_defaults and _character_payload_looks_default(config_manager, name, current_character):
-                continue
-            chosen = current_character
-        elif current_character is None:
-            chosen = legacy_character
-        else:
-            current_default = _character_payload_looks_default(config_manager, name, current_character)
-            legacy_default = _character_payload_looks_default(config_manager, name, legacy_character)
-            if current_default and not legacy_default:
-                chosen = legacy_character
-            elif legacy_default and not current_default:
-                chosen = current_character
-            else:
-                chosen = current_character
-        if chosen is not None:
-            merged_catgirls[name] = deepcopy(chosen)
-    merged_payload["猫娘"] = merged_catgirls
-
-    legacy_master = legacy_dict.get("主人")
-    current_master = current_dict.get("主人")
-    if legacy_master is None:
-        if current_master is not None:
-            merged_payload["主人"] = deepcopy(current_master)
-    elif current_master is None:
-        merged_payload["主人"] = deepcopy(legacy_master)
-    else:
-        current_master_default = _master_payload_looks_default(config_manager, current_master)
-        legacy_master_default = _master_payload_looks_default(config_manager, legacy_master)
-        chosen_master = legacy_master if current_master_default and not legacy_master_default else current_master
-        merged_payload["主人"] = deepcopy(chosen_master)
-
-    current_current_name = str(current_dict.get("当前猫娘") or "")
-    legacy_current_name = str(legacy_dict.get("当前猫娘") or "")
-    if current_current_name and current_current_name in merged_catgirls:
-        current_current_payload = current_catgirls.get(current_current_name)
-        current_default = _character_payload_looks_default(config_manager, current_current_name, current_current_payload)
-        if current_current_name not in legacy_catgirls and not preserve_current_only_defaults and current_default:
-            current_current_name = ""
-        elif current_current_name not in legacy_catgirls or not current_default:
-            merged_payload["当前猫娘"] = current_current_name
-        elif legacy_current_name and legacy_current_name in merged_catgirls:
-            merged_payload["当前猫娘"] = legacy_current_name
-        else:
-            merged_payload["当前猫娘"] = current_current_name
-    elif legacy_current_name and legacy_current_name in merged_catgirls:
-        merged_payload["当前猫娘"] = legacy_current_name
-    elif current_current_name and current_current_name in merged_catgirls:
-        merged_payload["当前猫娘"] = current_current_name
-    elif merged_catgirls:
-        merged_payload["当前猫娘"] = next(iter(merged_catgirls))
-    else:
-        merged_payload["当前猫娘"] = ""
-
-    return merged_payload
-
-
-def _runtime_root_summary(config_manager, root: Path) -> dict[str, Any]:
-    config_root = root / "config"
-    characters_path = config_root / "characters.json"
-    user_preferences_path = config_root / "user_preferences.json"
-    voice_storage_path = config_root / "voice_storage.json"
-    workshop_config_path = config_root / "workshop_config.json"
-    core_config_path = config_root / "core_config.json"
-
-    characters_payload = _load_json_if_exists(characters_path)
-    user_preferences_payload = _load_json_if_exists(user_preferences_path)
-    voice_storage_payload = _load_json_if_exists(voice_storage_path)
-    core_config_payload = _load_json_if_exists(core_config_path)
-    if not isinstance(characters_payload, dict):
-        characters_payload = None
-    character_names = set((characters_payload or {}).get("猫娘", {}) or {})
-    default_character_names = set((_load_seed_characters_payload(config_manager).get("猫娘") or {}).keys())
-
-    asset_dirs_with_content = {
-        dir_name: _directory_has_meaningful_content(root / dir_name)
-        for dir_name in RUNTIME_ASSET_DIR_NAMES
-    }
-    memory_character_names = _collect_memory_character_names(root)
-    seeded_character_shell = (
-        character_names.issubset(default_character_names)
-        and not memory_character_names
-        and not any(asset_dirs_with_content.values())
-    )
-    score = (
-        len(character_names) * 3
-        + len(memory_character_names) * 2
-        + (3 if user_preferences_path.is_file() else 0)
-        + (2 if voice_storage_path.is_file() else 0)
-        + (1 if workshop_config_path.is_file() else 0)
-        + (1 if core_config_path.is_file() else 0)
-        + sum(2 for has_content in asset_dirs_with_content.values() if has_content)
-    )
-
-    return {
-        "has_user_content": _runtime_root_has_user_content(root, config_manager=config_manager),
-        "characters_payload": characters_payload,
-        "character_names": character_names,
-        "memory_character_names": memory_character_names,
-        "has_user_preferences": user_preferences_path.is_file(),
-        "has_voice_storage": voice_storage_path.is_file(),
-        "has_workshop_config": workshop_config_path.is_file(),
-        "has_core_config": core_config_path.is_file(),
-        "asset_dirs_with_content": asset_dirs_with_content,
-        "seeded_character_shell": seeded_character_shell,
-        "looks_like_seeded": (
-            bool(character_names)
-            and character_names.issubset(default_character_names)
-            and not memory_character_names
-            and (
-                not user_preferences_path.is_file()
-                or _config_payload_looks_seeded(config_manager, "user_preferences.json", user_preferences_payload)
-            )
-            and (
-                not voice_storage_path.is_file()
-                or _config_payload_looks_seeded(config_manager, "voice_storage.json", voice_storage_payload)
-            )
-            and not workshop_config_path.is_file()
-            and (
-                not core_config_path.is_file()
-                or _config_payload_looks_seeded(config_manager, "core_config.json", core_config_payload)
-            )
-            and not any(asset_dirs_with_content.values())
-        ),
-        "score": score,
-    }
-
-
-def _legacy_root_provides_repair_benefit(config_manager, source_summary: dict[str, Any], target_summary: dict[str, Any]) -> tuple[bool, str]:
-    if not target_summary["has_user_content"]:
-        return True, "target_missing"
-
-    source_is_richer = source_summary["score"] > target_summary["score"]
-    target_is_seed_shell = bool(target_summary.get("seeded_character_shell"))
-
-    if target_is_seed_shell:
-        if source_summary["character_names"] - target_summary["character_names"]:
-            return True, "missing_characters"
-
-        if source_summary["memory_character_names"] - target_summary["memory_character_names"]:
-            return True, "missing_memory"
-
-        for flag_name, reason in (
-            ("has_user_preferences", "missing_user_preferences"),
-            ("has_voice_storage", "missing_voice_storage"),
-            ("has_workshop_config", "missing_workshop_config"),
-            ("has_core_config", "missing_core_config"),
-        ):
-            if source_summary[flag_name] and not target_summary[flag_name]:
-                return True, reason
-
-        for dir_name, source_has_content in source_summary["asset_dirs_with_content"].items():
-            if source_has_content and not target_summary["asset_dirs_with_content"].get(dir_name):
-                return True, f"missing_{dir_name}"
-
-    source_characters = (source_summary.get("characters_payload") or {}).get("猫娘", {}) or {}
-    target_characters = (target_summary.get("characters_payload") or {}).get("猫娘", {}) or {}
-    for name in sorted(set(source_characters) & set(target_characters)):
-        if (
-            _character_payload_looks_default(config_manager, name, target_characters.get(name))
-            and not _character_payload_looks_default(config_manager, name, source_characters.get(name))
-        ):
-            return True, "upgrade_default_character"
-
-    if target_is_seed_shell and source_is_richer:
-        return True, "repair_seeded_target"
-
-    return False, ""
-
-
-def _stage_merged_runtime_configs(config_manager, *, source_root: Path, target_root: Path, temp_root: Path, target_summary: dict[str, Any]) -> None:
-    config_dir = temp_root / "config"
-    config_dir.mkdir(parents=True, exist_ok=True)
-    target_tombstone_names = _load_tombstone_names_from_state_path(
-        target_root / "state" / "character_tombstones.json"
-    )
-
-    source_characters = _load_json_if_exists(source_root / "config" / "characters.json")
-    target_characters = _load_json_if_exists(target_root / "config" / "characters.json")
-    if source_characters is not None or target_characters is not None:
-        merged_characters = _merge_characters_payloads(
-            config_manager,
-            source_characters,
-            target_characters,
-            preserve_current_only_defaults=not bool(target_summary.get("seeded_character_shell")),
-        )
-        if target_tombstone_names:
-            merged_catgirls = merged_characters.get("猫娘") or {}
-            for deleted_name in target_tombstone_names:
-                merged_catgirls.pop(deleted_name, None)
-            merged_characters["猫娘"] = merged_catgirls
-            current_name = str(merged_characters.get("当前猫娘") or "")
-            if current_name in target_tombstone_names:
-                merged_characters["当前猫娘"] = next(iter(merged_catgirls), "")
-        atomic_write_json(config_dir / "characters.json", merged_characters, ensure_ascii=False, indent=2)
-
-    source_preferences = _load_json_if_exists(source_root / "config" / "user_preferences.json")
-    target_preferences = _load_json_if_exists(target_root / "config" / "user_preferences.json")
-    if source_preferences is not None or target_preferences is not None:
-        merged_preferences = _merge_preferences_payloads(source_preferences, target_preferences)
-        atomic_write_json(config_dir / "user_preferences.json", merged_preferences, ensure_ascii=False, indent=2)
-
-    for filename in ROOT_CONFIG_MERGE_FILES:
-        source_payload = _load_json_if_exists(source_root / "config" / filename)
-        target_payload = _load_json_if_exists(target_root / "config" / filename)
-        if source_payload is None and target_payload is None:
-            continue
-        merged_payload = _deep_merge_json_dicts(source_payload, target_payload)
-        if filename == "workshop_config.json":
-            merged_payload = rebase_runtime_bound_workshop_config_paths(
-                merged_payload,
-                source_root=source_root,
-                target_root=target_root,
-            )
-        atomic_write_json(config_dir / filename, merged_payload, ensure_ascii=False, indent=2)
-
-
-def _copy_optional_legacy_state(*, source_root: Path, target_root: Path, temp_root: Path) -> list[str]:
-    copied_paths: list[str] = []
-    for filename in TARGET_OPTIONAL_STATE_FILES:
-        target_path = target_root / "state" / filename
-        if not target_path.is_file():
-            continue
-        _stage_file_copy(temp_root, f"state/{filename}", target_path)
-        copied_paths.append(f"state/{filename}")
-    for filename in LEGACY_OPTIONAL_STATE_FILES:
-        source_path = source_root / "state" / filename
-        if not source_path.is_file() or (temp_root / "state" / filename).is_file():
-            continue
-        _stage_file_copy(temp_root, f"state/{filename}", source_path)
-        copied_paths.append(f"state/{filename}")
-    return copied_paths
-
-
-def _create_legacy_import_backup_path(target_root: Path) -> Path:
-    backup_pool = target_root.parent / f".{target_root.name}.legacy-import-backups"
-    backup_pool.mkdir(parents=True, exist_ok=True)
-    backup_slot = Path(tempfile.mkdtemp(prefix="backup-", dir=str(backup_pool)))
-    return backup_slot / target_root.name
-
-
-def _replace_runtime_root(target_root: Path, temp_root: Path, *, backup_path: Path | None = None) -> None:
-    if backup_path is None:
-        if target_root.exists():
-            shutil.rmtree(target_root, ignore_errors=True)
-        os.replace(temp_root, target_root)
-        return
-
-    restore_required = False
-    try:
-        if target_root.exists():
-            os.replace(target_root, backup_path)
-            restore_required = True
-        os.replace(temp_root, target_root)
-    except Exception:
-        if restore_required and backup_path.exists() and not target_root.exists():
-            os.replace(backup_path, target_root)
-        raise
-
-
-def _legacy_source_was_already_imported(
-    root_state: Any,
-    *,
-    source_root: Path,
-    target_root: Path,
-) -> bool:
-    """Treat legacy root import as a one-shot bootstrap repair per source root.
-
-    Once a legacy root has already been imported and the migrated target has
-    completed at least one successful boot, future startups should treat the
-    current runtime root as the source of truth. Otherwise, deletions performed
-    in the new runtime root can be "repaired" back from the stale legacy root.
-    """
-    if not isinstance(root_state, dict):
-        return False
-    if str(root_state.get("current_root") or "") != str(target_root):
-        return False
-    if not str(root_state.get("last_successful_boot_at") or "").strip():
-        return False
-    if str(root_state.get("last_migration_source") or "") != str(source_root):
-        return False
-    last_result = str(root_state.get("last_migration_result") or "")
-    return last_result.startswith("legacy_root_")
-
-
-def _root_has_staged_cloudsave_snapshot(root: Path) -> bool:
-    manifest_path = Path(root) / "cloudsave" / "manifest.json"
-    if manifest_path.is_file():
-        try:
-            manifest_payload = json.loads(manifest_path.read_text(encoding="utf-8"))
-            manifest_files = manifest_payload.get("files")
-            if isinstance(manifest_files, dict) and manifest_files:
-                return True
-        except Exception:
-            pass
-
-    cloudsave_root = Path(root) / "cloudsave"
-    if not cloudsave_root.exists():
-        return False
-    try:
-        for child in cloudsave_root.rglob("*"):
-            if child.is_file() and child.name != "manifest.json":
-                return True
-    except Exception:
-        return False
-    return False
-
-
-def import_legacy_runtime_root_if_needed(config_manager) -> dict[str, Any]:
-    """One-time bootstrap import from legacy roots into the deterministic app data root."""
-    target_root = Path(config_manager.app_docs_dir)
-    target_has_user_content = _runtime_root_has_user_content(target_root, config_manager=config_manager)
-    target_has_staged_cloudsave_snapshot = _root_has_staged_cloudsave_snapshot(target_root)
-    target_summary = _runtime_root_summary(config_manager, target_root)
-    existing_root_state = None
-    try:
-        if config_manager.root_state_path.is_file():
-            existing_root_state = config_manager.load_root_state()
-    except Exception:
-        existing_root_state = None
-
-    if target_has_staged_cloudsave_snapshot and not target_has_user_content:
-        return {
-            "migrated": False,
-            "source": "",
-            "copied_paths": [],
-            "backup_path": "",
-            "repair_reason": "",
-            "result": "target_root_preserves_staged_cloudsave_snapshot",
-        }
-
-    saw_legacy_source = False
-
-    for source_root in config_manager.get_legacy_app_root_candidates():
-        source_root = Path(source_root)
-        if not _runtime_root_has_user_content(source_root, config_manager=config_manager):
-            continue
-        saw_legacy_source = True
-        if _legacy_source_was_already_imported(
-            existing_root_state,
-            source_root=source_root,
-            target_root=target_root,
-        ):
-            continue
-
-        source_summary = _runtime_root_summary(config_manager, source_root)
-        should_repair, repair_reason = _legacy_root_provides_repair_benefit(
-            config_manager,
-            source_summary,
-            target_summary,
-        )
-        if target_has_user_content and not should_repair:
-            continue
-
-        temp_root = target_root.parent / f".{target_root.name}.bootstrap-import"
-        if temp_root.exists():
-            shutil.rmtree(temp_root, ignore_errors=True)
-        temp_root.parent.mkdir(parents=True, exist_ok=True)
-
-        copied_paths: list[str] = []
-        backup_path: Path | None = None
-        try:
-            copied_paths.extend(_copy_runtime_root_entries(source_root, temp_root))
-            if target_has_user_content:
-                _copy_runtime_root_entries(target_root, temp_root)
-                _stage_merged_runtime_configs(
-                    config_manager,
-                    source_root=source_root,
-                    target_root=target_root,
-                    temp_root=temp_root,
-                    target_summary=target_summary,
-                )
-                backup_path = _create_legacy_import_backup_path(target_root)
-            copied_paths.extend(_copy_optional_legacy_state(source_root=source_root, target_root=target_root, temp_root=temp_root))
-
-            if not copied_paths:
-                shutil.rmtree(temp_root, ignore_errors=True)
-                continue
-
-            _replace_runtime_root(target_root, temp_root, backup_path=backup_path)
-            return {
-                "migrated": True,
-                "source": str(source_root),
-                "copied_paths": sorted(set(copied_paths)),
-                "backup_path": str(backup_path) if backup_path is not None else "",
-                "repair_reason": repair_reason,
-                "result": "legacy_root_repaired_target" if target_has_user_content else "legacy_root_imported",
-            }
-        except Exception:
-            shutil.rmtree(temp_root, ignore_errors=True)
-            raise
-
-    if target_has_user_content:
-        return {
-            "migrated": False,
-            "source": "",
-            "copied_paths": [],
-            "backup_path": "",
-            "repair_reason": "",
-            "result": "target_root_already_initialized" if saw_legacy_source or target_summary["has_user_content"] else "no_legacy_root_found",
-        }
-
-    return {
-        "migrated": False,
-        "source": "",
-        "copied_paths": [],
-        "backup_path": "",
-        "repair_reason": "",
-        "result": "no_legacy_root_found",
-    }
-
-
-def _load_user_preferences_entries(config_manager) -> list[dict[str, Any]]:
-    preferences_path = Path(config_manager.get_config_path("user_preferences.json"))
-    if not preferences_path.exists():
-        return []
-    try:
-        with open(preferences_path, "r", encoding="utf-8") as file_obj:
-            payload = json.load(file_obj)
-    except Exception:
-        return []
-    if isinstance(payload, dict):
-        return [payload]
-    if isinstance(payload, list):
-        return payload
-    return []
-
-
-def _extract_conversation_settings(config_manager) -> dict[str, Any]:
-    for entry in _load_user_preferences_entries(config_manager):
-        if isinstance(entry, dict) and entry.get("model_path") == GLOBAL_CONVERSATION_KEY:
-            return {
-                key: value
-                for key, value in entry.items()
-                if key != "model_path"
-            }
-    return {}
-
-
-def _build_runtime_preferences_payload(config_manager, conversation_settings: dict[str, Any]) -> list[dict[str, Any]]:
-    preferences = [
-        entry
-        for entry in _load_user_preferences_entries(config_manager)
-        if not isinstance(entry, dict) or entry.get("model_path") != GLOBAL_CONVERSATION_KEY
-    ]
-    filtered_settings = {
-        key: value
-        for key, value in (conversation_settings or {}).items()
-        if key != "model_path"
-    }
-    if filtered_settings:
-        preferences.append({
-            "model_path": GLOBAL_CONVERSATION_KEY,
-            **filtered_settings,
-        })
-    return preferences
-
-
-def _derive_binding_model_reference(character_payload: dict[str, Any]) -> tuple[str, str]:
-    from utils.config_manager import get_reserved
-
-    runtime_model_type = str(
-        get_reserved(character_payload, "avatar", "model_type", default="live2d", legacy_keys=("model_type",))
-    ).strip().lower()
-    live2d_model_path = str(
-        get_reserved(character_payload, "avatar", "live2d", "model_path", default="", legacy_keys=("live2d",))
-        or ""
-    ).strip()
-    vrm_model_path = str(
-        get_reserved(character_payload, "avatar", "vrm", "model_path", default="", legacy_keys=("vrm",))
-        or ""
-    ).strip()
-    mmd_model_path = str(
-        get_reserved(character_payload, "avatar", "mmd", "model_path", default="")
-        or ""
-    ).strip()
-
-    if runtime_model_type in {"live3d", "vrm"}:
-        if mmd_model_path:
-            return "mmd", mmd_model_path.replace("\\", "/")
-        if vrm_model_path:
-            return "vrm", vrm_model_path.replace("\\", "/")
-        if live2d_model_path:
-            return "live2d", live2d_model_path.replace("\\", "/")
-        return "vrm", ""
-
-    return "live2d", live2d_model_path.replace("\\", "/")
-
-
-def _derive_binding_asset_source(*, model_ref: str, stored_asset_source: str, asset_source_id: str) -> str:
-    normalized_ref = str(model_ref or "").strip().replace("\\", "/")
-    normalized_source = str(stored_asset_source or "").strip().lower()
-
-    if normalized_source == "steam_workshop" or asset_source_id or normalized_ref.startswith("/workshop/"):
-        return "steam_workshop"
-    if normalized_source == "builtin":
-        return "builtin"
-    if normalized_source in {"manual_external", "external"}:
-        return "manual_external"
-    if normalized_source in {"local_imported", "local"}:
-        return "local_imported"
-    if normalized_ref.startswith(("http://", "https://")):
-        return "manual_external"
-    if normalized_ref.startswith(("/user_live2d/", "/user_live2d_local/", "/user_vrm/", "/user_mmd/")):
-        return "local_imported"
-    if normalized_ref.startswith("/static/") or (normalized_ref and not normalized_ref.startswith("/")):
-        return "builtin"
-    return "local_imported" if normalized_ref else ""
-
-
-def _derive_binding_asset_source_id(*, model_ref: str, stored_source_id: str) -> str:
-    source_id = str(stored_source_id or "").strip()
-    if source_id:
-        return source_id
-    normalized_ref = str(model_ref or "").strip().replace("\\", "/")
-    if normalized_ref.startswith("/workshop/"):
-        parts = normalized_ref.split("/")
-        if len(parts) >= 3:
-            return parts[2]
-    return ""
-
-
-def _derive_binding_asset_display_name(model_ref: str) -> str:
-    normalized_ref = str(model_ref or "").strip().replace("\\", "/")
-    if not normalized_ref:
-        return ""
-    if normalized_ref.endswith(".model3.json"):
-        parts = [part for part in normalized_ref.split("/") if part]
-        if len(parts) >= 2:
-            return parts[-2]
-        return Path(parts[-1]).stem.replace(".model3", "")
-    if normalized_ref.endswith((".vrm", ".pmx", ".pmd", ".vmd", ".vrma")):
-        return Path(normalized_ref).stem
-    parts = [part for part in normalized_ref.split("/") if part]
-    return parts[-1] if parts else normalized_ref
-
-
-def _collect_binding_live2d_roots(config_manager) -> list[Path]:
-    get_live2d_lookup_roots = getattr(config_manager, "get_live2d_lookup_roots", None)
-    if callable(get_live2d_lookup_roots):
-        try:
-            return [Path(candidate) for candidate in get_live2d_lookup_roots(prefer_writable=True)]
-        except Exception:
-            pass
-
-    roots: list[Path] = []
-    seen_roots: set[str] = set()
-    for candidate in (
-        getattr(config_manager, "live2d_dir", None),
-        getattr(config_manager, "readable_live2d_dir", None),
-    ):
-        if not candidate:
-            continue
-        normalized_root = os.path.normcase(os.path.normpath(str(candidate)))
-        if normalized_root in seen_roots:
-            continue
-        seen_roots.add(normalized_root)
-        roots.append(Path(candidate))
-    return roots
-
-
-def _collect_binding_workshop_roots(config_manager) -> list[Path]:
-    roots: list[Path] = []
-    seen_roots: set[str] = set()
-
-    get_workshop_path = getattr(config_manager, "get_workshop_path", None)
-    if callable(get_workshop_path):
-        try:
-            configured_workshop_root = get_workshop_path()
-        except Exception:
-            configured_workshop_root = ""
-        if configured_workshop_root:
-            normalized_root = os.path.normcase(os.path.normpath(str(configured_workshop_root)))
-            if normalized_root not in seen_roots:
-                seen_roots.add(normalized_root)
-                roots.append(Path(configured_workshop_root))
-
-    fallback_workshop_root = getattr(config_manager, "workshop_dir", None)
-    if fallback_workshop_root:
-        normalized_root = os.path.normcase(os.path.normpath(str(fallback_workshop_root)))
-        if normalized_root not in seen_roots:
-            seen_roots.add(normalized_root)
-            roots.append(Path(fallback_workshop_root))
-
-    return roots
-
-
-def _normalize_workshop_character_model_ref(model_type: str, payload: dict[str, Any]) -> str:
-    normalized_type = str(model_type or "").strip().lower()
-    if normalized_type != "live2d":
-        return ""
-
-    live2d_name = str(payload.get("live2d") or "").strip().replace("\\", "/")
-    if not live2d_name:
-        return ""
-    if live2d_name.endswith(".model3.json") or "/" in live2d_name:
-        return live2d_name
-    return f"{live2d_name}/{live2d_name}.model3.json"
-
-
-def _build_character_origin_match_payload(payload: Any) -> dict[str, Any]:
-    normalized_payload = _normalize_catgirl_payload(payload)
-    if normalized_payload is None:
-        return {}
-
-    skip_keys = {"档案名", *CHARACTER_RESERVED_FIELDS}
-    comparable_payload: dict[str, Any] = {}
-    for key, value in normalized_payload.items():
-        if key in skip_keys or value is None:
-            continue
-        comparable_payload[key] = deepcopy(value)
-    return comparable_payload
-
-
-def _build_character_origin_profile_fingerprint(payload: Any) -> str:
-    comparable_payload = _build_character_origin_match_payload(payload)
-    if not comparable_payload:
-        return ""
-
-    fingerprint_payload = {
-        "schema_version": 1,
-        "character_payload": comparable_payload,
-    }
-    return "sha256:" + _sha256_bytes(_json_canonical_dumps(fingerprint_payload).encode("utf-8"))
-
-
-def _collect_workshop_character_origin_candidates(config_manager) -> dict[str, list[dict[str, Any]]]:
-    candidates_by_name: dict[str, list[dict[str, Any]]] = {}
-    seen_entries: set[tuple[str, str, str, str, str]] = set()
-
-    for workshop_root in _collect_binding_workshop_roots(config_manager):
-        if not workshop_root.is_dir():
-            continue
-        try:
-            item_roots = sorted(child for child in workshop_root.iterdir() if child.is_dir())
-        except Exception:
-            continue
-
-        for item_root in item_roots:
-            item_id = str(item_root.name or "").strip()
-            if not item_id:
-                continue
-            try:
-                chara_paths = sorted(path for path in item_root.rglob("*.chara.json") if path.is_file())
-            except Exception:
-                continue
-
-            for chara_path in chara_paths:
-                payload = _load_json_if_exists(chara_path)
-                if not isinstance(payload, dict):
-                    continue
-
-                character_name = str(payload.get("档案名") or payload.get("name") or "").strip()
-                if not character_name:
-                    continue
-
-                model_type = str(payload.get("model_type") or "live2d").strip().lower() or "live2d"
-                model_ref = _normalize_workshop_character_model_ref(model_type, payload)
-                origin_profile_fingerprint = _build_character_origin_profile_fingerprint(payload)
-                dedupe_key = (character_name, item_id, model_type, model_ref, origin_profile_fingerprint)
-                if dedupe_key in seen_entries:
-                    continue
-                seen_entries.add(dedupe_key)
-
-                candidates_by_name.setdefault(character_name, []).append(
-                    {
-                        "character_name": character_name,
-                        "origin_source": "steam_workshop",
-                        "origin_source_id": item_id,
-                        "model_type": model_type,
-                        "origin_model_ref": model_ref,
-                        "origin_display_name": _derive_binding_asset_display_name(model_ref),
-                        "origin_profile_fingerprint": origin_profile_fingerprint,
-                    }
-                )
-
-    return candidates_by_name
-
-
-def _select_workshop_character_origin_candidate(
-    candidates: list[dict[str, Any]],
-    *,
-    model_type: str,
-    origin_source_id_hint: str = "",
-    origin_model_ref_hint: str = "",
-    origin_profile_fingerprint_hint: str = "",
-) -> dict[str, Any] | None:
-    if not candidates:
-        return None
-
-    selected_pool = [
-        candidate
-        for candidate in candidates
-        if not candidate.get("model_type") or str(candidate.get("model_type") or "") == str(model_type or "")
-    ] or list(candidates)
-
-    origin_source_id_hint = str(origin_source_id_hint or "").strip()
-    if origin_source_id_hint:
-        id_matches = [
-            candidate
-            for candidate in selected_pool
-            if str(candidate.get("origin_source_id") or "").strip() == origin_source_id_hint
-        ]
-        if len(id_matches) == 1:
-            return deepcopy(id_matches[0])
-        if id_matches:
-            selected_pool = id_matches
-
-    origin_model_ref_hint = str(origin_model_ref_hint or "").strip().replace("\\", "/")
-    if origin_model_ref_hint:
-        exact_ref_matches = [
-            candidate
-            for candidate in selected_pool
-            if str(candidate.get("origin_model_ref") or "").strip().replace("\\", "/") == origin_model_ref_hint
-        ]
-        if len(exact_ref_matches) == 1:
-            return deepcopy(exact_ref_matches[0])
-        if exact_ref_matches:
-            selected_pool = exact_ref_matches
-
-    origin_profile_fingerprint_hint = str(origin_profile_fingerprint_hint or "").strip()
-    if origin_profile_fingerprint_hint:
-        fingerprint_matches = [
-            candidate
-            for candidate in selected_pool
-            if str(candidate.get("origin_profile_fingerprint") or "").strip() == origin_profile_fingerprint_hint
-        ]
-        if len(fingerprint_matches) == 1:
-            return deepcopy(fingerprint_matches[0])
-
-    return None
-
-
-def _derive_character_origin_metadata(
-    config_manager,
-    *,
-    character_name: str,
-    character_payload: dict[str, Any],
-    model_type: str,
-    workshop_origin_index: dict[str, list[dict[str, Any]]] | None = None,
-) -> dict[str, str]:
-    from utils.config_manager import get_reserved
-
-    origin_source = ""
-    origin_source_id = ""
-    origin_model_ref = ""
-    origin_display_name = ""
-
-    origin_source = str(get_reserved(character_payload, "character_origin", "source", default="") or "").strip()
-    origin_source_id = str(get_reserved(character_payload, "character_origin", "source_id", default="") or "").strip()
-    origin_model_ref = str(get_reserved(character_payload, "character_origin", "model_ref", default="") or "").strip().replace("\\", "/")
-    origin_display_name = str(get_reserved(character_payload, "character_origin", "display_name", default="") or "").strip()
-    origin_profile_fingerprint = _build_character_origin_profile_fingerprint(character_payload)
-
-    candidates = (workshop_origin_index or {}).get(character_name) or []
-    selected_candidate = _select_workshop_character_origin_candidate(
-        candidates,
-        model_type=model_type,
-        origin_source_id_hint=origin_source_id,
-        origin_model_ref_hint=origin_model_ref,
-        origin_profile_fingerprint_hint=origin_profile_fingerprint,
-    )
-    if selected_candidate is not None:
-        if not origin_source:
-            origin_source = str(selected_candidate.get("origin_source") or "")
-        if not origin_source_id:
-            origin_source_id = str(selected_candidate.get("origin_source_id") or "")
-        if not origin_model_ref:
-            origin_model_ref = str(selected_candidate.get("origin_model_ref") or "")
-        if not origin_display_name:
-            origin_display_name = str(selected_candidate.get("origin_display_name") or "")
-
-    return {
-        "origin_source": origin_source,
-        "origin_source_id": origin_source_id,
-        "origin_model_ref": origin_model_ref,
-        "origin_display_name": origin_display_name,
-    }
-
-
-def _build_live2d_model_ref_hints(model_ref: str) -> tuple[str, str, str, str]:
-    normalized_ref = str(model_ref or "").strip().replace("\\", "/")
-    normalized_suffix = normalized_ref.lstrip("/")
-    if normalized_ref.startswith("/workshop/"):
-        parts = [part for part in normalized_ref.split("/") if part]
-        if len(parts) >= 3:
-            normalized_suffix = "/".join(parts[2:])
-
-    relative_parent = ""
-    if normalized_suffix:
-        relative_parent = Path(normalized_suffix).parent.as_posix()
-        if relative_parent == ".":
-            relative_parent = ""
-
-    expected_filename = Path(normalized_ref).name if normalized_ref else ""
-    expected_folder_name = ""
-    expected_model_name = ""
-    if normalized_ref.endswith(".model3.json"):
-        parts = [part for part in normalized_ref.split("/") if part]
-        if len(parts) >= 2:
-            expected_folder_name = parts[-2]
-        expected_model_name = Path(expected_filename).stem.replace(".model3", "")
-    elif normalized_ref:
-        expected_model_name = Path(normalized_ref).stem
-
-    return normalized_suffix, relative_parent, expected_filename, expected_folder_name or expected_model_name
-
-
-def _rank_live2d_model3_path(
-    candidate_path: Path,
-    *,
-    candidate_root: Path,
-    normalized_suffix: str,
-    relative_parent: str,
-    expected_filename: str,
-    expected_folder_name: str,
-) -> tuple[int, int, str, Path]:
-    try:
-        relative_path = candidate_path.relative_to(candidate_root).as_posix()
-    except Exception:
-        relative_path = candidate_path.name
-
-    expected_model_name = Path(expected_filename).stem.replace(".model3", "") if expected_filename else ""
-    candidate_model_name = candidate_path.stem.replace(".model3", "")
-
-    score = 0
-    if normalized_suffix and relative_path == normalized_suffix:
-        score += 100
-    elif normalized_suffix and relative_path.endswith(normalized_suffix):
-        score += 80
-    if relative_parent and relative_path.startswith(f"{relative_parent}/"):
-        score += 20
-    if expected_filename and candidate_path.name == expected_filename:
-        score += 40
-    if expected_folder_name and candidate_path.parent.name == expected_folder_name:
-        score += 20
-    if expected_model_name and candidate_model_name == expected_model_name:
-        score += 10
-
-    return (score, -len(relative_path.split("/")), relative_path, candidate_path)
-
-
-def _is_path_within(candidate_path: Path, base_path: Path) -> bool:
-    try:
-        candidate_real = os.path.normcase(os.path.realpath(str(candidate_path)))
-        base_real = os.path.normcase(os.path.realpath(str(base_path)))
-        return os.path.commonpath([candidate_real, base_real]) == base_real
-    except Exception:
-        return False
-
-
-def _infer_binding_source_from_resolved_path(
-    config_manager,
-    *,
-    resolved_path: Path | None,
-    asset_source: str,
-    asset_source_id: str,
-) -> tuple[str, str]:
-    if resolved_path is None or not resolved_path.is_file():
-        return asset_source, asset_source_id
-
-    for workshop_root in _collect_binding_workshop_roots(config_manager):
-        if not _is_path_within(resolved_path, workshop_root):
-            continue
-        inferred_source_id = str(asset_source_id or "").strip()
-        if not inferred_source_id:
-            try:
-                relative_parts = resolved_path.relative_to(workshop_root).parts
-            except Exception:
-                relative_parts = ()
-            if relative_parts:
-                inferred_source_id = str(relative_parts[0])
-        return "steam_workshop", inferred_source_id
-
-    for live2d_root in _collect_binding_live2d_roots(config_manager):
-        if _is_path_within(resolved_path, live2d_root):
-            return "local_imported", ""
-
-    static_root = Path(config_manager.project_root) / "static"
-    if _is_path_within(resolved_path, static_root):
-        return "builtin", ""
-
-    return asset_source, asset_source_id
-
-
-def _resolve_binding_file_path(
-    config_manager,
-    *,
-    model_type: str,
-    model_ref: str,
-    asset_source: str,
-    asset_source_id: str,
-) -> Path | None:
-    normalized_ref = str(model_ref or "").strip().replace("\\", "/")
-    if not normalized_ref or normalized_ref.startswith(("http://", "https://")):
-        return None
-
-    candidates: list[Path] = []
-    live2d_roots = _collect_binding_live2d_roots(config_manager)
-    readable_live2d_dir = getattr(config_manager, "readable_live2d_dir", None)
-    workshop_roots = _collect_binding_workshop_roots(config_manager)
-
-    def _resolve_workshop_live2d_fallback() -> Path | None:
-        if model_type != "live2d" or asset_source != "steam_workshop" or not asset_source_id:
-            return None
-
-        normalized_suffix, relative_parent, expected_filename, expected_folder_name = _build_live2d_model_ref_hints(normalized_ref)
-
-        ranked_candidates: list[tuple[int, int, str, Path]] = []
-        for workshop_root in workshop_roots:
-            item_root = workshop_root / asset_source_id
-            if not item_root.is_dir():
-                continue
-            try:
-                discovered_files = sorted(path for path in item_root.rglob("*.model3.json") if path.is_file())
-            except Exception:
-                continue
-
-            for discovered_path in discovered_files:
-                try:
-                    relative_path = discovered_path.relative_to(item_root).as_posix()
-                except Exception:
-                    relative_path = discovered_path.name
-
-                ranked_candidates.append(
-                    _rank_live2d_model3_path(
-                        discovered_path,
-                        candidate_root=item_root,
-                        normalized_suffix=normalized_suffix,
-                        relative_parent=relative_parent,
-                        expected_filename=expected_filename,
-                        expected_folder_name=expected_folder_name,
-                    )
-                )
-
-        if not ranked_candidates:
-            return None
-
-        ranked_candidates.sort(key=lambda entry: (entry[0], entry[1], entry[2]), reverse=True)
-        return ranked_candidates[0][3]
-
-    def _resolve_local_live2d_fallback() -> Path | None:
-        if model_type != "live2d" or normalized_ref.startswith("/"):
-            return None
-
-        normalized_suffix, relative_parent, expected_filename, expected_folder_name = _build_live2d_model_ref_hints(normalized_ref)
-        ranked_candidates: list[tuple[int, int, str, Path]] = []
-
-        def _append_candidates(search_root: Path, candidate_base: Path) -> None:
-            if not candidate_base.is_dir():
-                return
-            try:
-                discovered_files = sorted(path for path in candidate_base.rglob("*.model3.json") if path.is_file())
-            except Exception:
-                return
-            for discovered_path in discovered_files:
-                ranked_candidates.append(
-                    _rank_live2d_model3_path(
-                        discovered_path,
-                        candidate_root=search_root,
-                        normalized_suffix=normalized_suffix,
-                        relative_parent=relative_parent,
-                        expected_filename=expected_filename,
-                        expected_folder_name=expected_folder_name,
-                    )
-                )
-
-        for live2d_root in live2d_roots:
-            candidate_dirs: list[Path] = []
-            if relative_parent:
-                candidate_dirs.append(live2d_root / relative_parent)
-            if expected_folder_name:
-                candidate_dirs.append(live2d_root / expected_folder_name)
-
-            seen_candidate_dirs: set[str] = set()
-            for candidate_dir in candidate_dirs:
-                normalized_dir = os.path.normcase(os.path.normpath(str(candidate_dir)))
-                if normalized_dir in seen_candidate_dirs:
-                    continue
-                seen_candidate_dirs.add(normalized_dir)
-                _append_candidates(live2d_root, candidate_dir)
-
-        for workshop_root in workshop_roots:
-            try:
-                item_roots = sorted(child for child in workshop_root.iterdir() if child.is_dir())
-            except Exception:
-                continue
-            for item_root in item_roots:
-                candidate_dirs: list[Path] = []
-                if relative_parent:
-                    candidate_dirs.append(item_root / relative_parent)
-                if expected_folder_name:
-                    candidate_dirs.append(item_root / expected_folder_name)
-                if not candidate_dirs:
-                    candidate_dirs.append(item_root)
-
-                seen_candidate_dirs: set[str] = set()
-                for candidate_dir in candidate_dirs:
-                    normalized_dir = os.path.normcase(os.path.normpath(str(candidate_dir)))
-                    if normalized_dir in seen_candidate_dirs:
-                        continue
-                    seen_candidate_dirs.add(normalized_dir)
-                    _append_candidates(item_root, candidate_dir)
-
-        if not ranked_candidates:
-            return None
-
-        ranked_candidates.sort(key=lambda entry: (entry[0], entry[1], entry[2]), reverse=True)
-        return ranked_candidates[0][3]
-
-    if model_type == "live2d":
-        if normalized_ref.startswith("/user_live2d/"):
-            relative_part = normalized_ref[len("/user_live2d/"):]
-            if readable_live2d_dir is not None:
-                candidates.append(Path(readable_live2d_dir) / relative_part)
-            candidates.append(Path(config_manager.live2d_dir) / relative_part)
-        elif normalized_ref.startswith("/user_live2d_local/"):
-            candidates.append(Path(config_manager.live2d_dir) / normalized_ref[len("/user_live2d_local/"):])
-        elif normalized_ref.startswith("/workshop/"):
-            relative_workshop_path = "/".join(normalized_ref.split("/")[2:])
-            for workshop_root in workshop_roots:
-                candidates.append(workshop_root / relative_workshop_path)
-        else:
-            if asset_source == "steam_workshop" and asset_source_id:
-                for workshop_root in workshop_roots:
-                    candidates.append(workshop_root / asset_source_id / normalized_ref)
-                    candidates.append(workshop_root / asset_source_id / Path(normalized_ref).name)
-            if asset_source == "local_imported":
-                if readable_live2d_dir is not None:
-                    candidates.append(Path(readable_live2d_dir) / normalized_ref)
-                candidates.append(Path(config_manager.live2d_dir) / normalized_ref)
-            candidates.append(Path(config_manager.project_root) / "static" / normalized_ref)
-    elif model_type == "vrm":
-        if normalized_ref.startswith("/user_vrm/"):
-            candidates.append(Path(config_manager.vrm_dir) / normalized_ref[len("/user_vrm/"):])
-        elif normalized_ref.startswith("/static/vrm/"):
-            candidates.append(Path(config_manager.project_root) / "static" / "vrm" / normalized_ref[len("/static/vrm/"):])
-        elif normalized_ref.startswith("/workshop/"):
-            relative_workshop_path = "/".join(normalized_ref.split("/")[2:])
-            for workshop_root in workshop_roots:
-                candidates.append(workshop_root / relative_workshop_path)
-    elif model_type == "mmd":
-        if normalized_ref.startswith("/user_mmd/"):
-            candidates.append(Path(config_manager.mmd_dir) / normalized_ref[len("/user_mmd/"):])
-        elif normalized_ref.startswith("/static/mmd/"):
-            candidates.append(Path(config_manager.project_root) / "static" / "mmd" / normalized_ref[len("/static/mmd/"):])
-        elif normalized_ref.startswith("/workshop/"):
-            relative_workshop_path = "/".join(normalized_ref.split("/")[2:])
-            for workshop_root in workshop_roots:
-                candidates.append(workshop_root / relative_workshop_path)
-
-    for candidate in candidates:
-        if candidate.is_file():
-            return candidate
-
-    fallback_candidate = _resolve_workshop_live2d_fallback()
-    if fallback_candidate is not None and fallback_candidate.is_file():
-        return fallback_candidate
-    fallback_candidate = _resolve_local_live2d_fallback()
-    if fallback_candidate is not None and fallback_candidate.is_file():
-        return fallback_candidate
-    return None
-
-
-def _derive_binding_asset_state(*, resolved_path: Path | None, asset_source: str, model_ref: str) -> str:
-    if resolved_path is not None and resolved_path.is_file():
-        return "ready"
-    if not str(model_ref or "").strip():
-        return "missing"
-    if asset_source == "steam_workshop":
-        return "downloadable"
-    if asset_source in {"local_imported", "manual_external"}:
-        return "import_required"
-    return "missing"
-
-
-def _derive_binding_experience_overrides(character_payload: dict[str, Any]) -> dict[str, Any]:
-    from utils.config_manager import get_reserved
-
-    overrides = {
-        "touch_set": deepcopy(get_reserved(character_payload, "touch_set", default={}) or {}),
-        "vrm_lighting": deepcopy(get_reserved(character_payload, "avatar", "vrm", "lighting", default={}) or {}),
-        "mmd_lighting": deepcopy(get_reserved(character_payload, "avatar", "mmd", "lighting", default={}) or {}),
-        "mmd_rendering": deepcopy(get_reserved(character_payload, "avatar", "mmd", "rendering", default={}) or {}),
-        "mmd_physics": deepcopy(get_reserved(character_payload, "avatar", "mmd", "physics", default={}) or {}),
-        "mmd_cursor_follow": deepcopy(get_reserved(character_payload, "avatar", "mmd", "cursor_follow", default={}) or {}),
-    }
-    return {
-        key: value
-        for key, value in overrides.items()
-        if value not in ({}, None, [])
-    }
-
-
-def _derive_character_binding_summary(
-    config_manager,
-    character_name: str,
-    character_payload: dict[str, Any],
-    *,
-    workshop_origin_index: dict[str, list[dict[str, Any]]] | None = None,
-) -> dict[str, Any]:
-    from utils.config_manager import get_reserved
-
-    binding_model_type, model_ref = _derive_binding_model_reference(character_payload)
-    stored_source = get_reserved(character_payload, "avatar", "asset_source", default="")
-    stored_source_id = get_reserved(
-        character_payload,
-        "avatar",
-        "asset_source_id",
-        default="",
-        legacy_keys=("live2d_item_id", "item_id"),
-    )
-    asset_source_id = _derive_binding_asset_source_id(model_ref=model_ref, stored_source_id=str(stored_source_id or ""))
-    asset_source = _derive_binding_asset_source(
-        model_ref=model_ref,
-        stored_asset_source=str(stored_source or ""),
-        asset_source_id=asset_source_id,
-    )
-    resolved_path = _resolve_binding_file_path(
-        config_manager,
-        model_type=binding_model_type,
-        model_ref=model_ref,
-        asset_source=asset_source,
-        asset_source_id=asset_source_id,
-    )
-    asset_source, asset_source_id = _infer_binding_source_from_resolved_path(
-        config_manager,
-        resolved_path=resolved_path,
-        asset_source=asset_source,
-        asset_source_id=asset_source_id,
-    )
-    asset_state = _derive_binding_asset_state(
-        resolved_path=resolved_path,
-        asset_source=asset_source,
-        model_ref=model_ref,
-    )
-    origin_payload = _derive_character_origin_metadata(
-        config_manager,
-        character_name=character_name,
-        character_payload=character_payload,
-        model_type=binding_model_type,
-        workshop_origin_index=workshop_origin_index,
-    )
-    asset_fingerprint = _sha256_file(resolved_path) if resolved_path is not None else ""
-
-    fallback_model_ref = ""
-    if asset_state != "ready" and binding_model_type != "live2d":
-        fallback_model_ref = "yui-origin/yui-origin.model3.json"
-
-    return {
-        "character_name": character_name,
-        "model_type": binding_model_type,
-        "asset_source": asset_source,
-        "asset_source_id": asset_source_id,
-        "model_ref": model_ref,
-        "asset_display_name": _derive_binding_asset_display_name(model_ref),
-        "asset_fingerprint": asset_fingerprint,
-        "asset_state": asset_state,
-        "origin_source": str(origin_payload.get("origin_source") or ""),
-        "origin_source_id": str(origin_payload.get("origin_source_id") or ""),
-        "origin_model_ref": str(origin_payload.get("origin_model_ref") or ""),
-        "origin_display_name": str(origin_payload.get("origin_display_name") or ""),
-        "fallback_model_ref": fallback_model_ref,
-        "last_verified_at": _utc_now_iso() if resolved_path is not None else "",
-        "experience_overrides": _derive_binding_experience_overrides(character_payload),
-    }
-
-
-def _build_catalog_index_payload(
-    *,
-    character_names: list[str],
-    characters_payload: dict[str, Any],
-    binding_payloads: dict[str, dict[str, Any]],
-    sequence_number: int,
-    exported_at: str,
-) -> dict[str, Any]:
-    catgirls_payload = characters_payload.get("猫娘") or {}
-    return {
-        "schema_version": 1,
-        "sequence_number": sequence_number,
-        "exported_at_utc": exported_at,
-        "characters": [
-            {
-                "character_name": name,
-                "entry_sequence_number": sequence_number,
-                "has_memory": True,
-                "model_type": binding_payloads.get(name, {}).get("model_type", ""),
-                "asset_source": binding_payloads.get(name, {}).get("asset_source", ""),
-                "asset_source_id": binding_payloads.get(name, {}).get("asset_source_id", ""),
-                "asset_state": binding_payloads.get(name, {}).get("asset_state", ""),
-                "origin_source": binding_payloads.get(name, {}).get("origin_source", ""),
-                "origin_source_id": binding_payloads.get(name, {}).get("origin_source_id", ""),
-                "origin_model_ref": binding_payloads.get(name, {}).get("origin_model_ref", ""),
-                "origin_display_name": binding_payloads.get(name, {}).get("origin_display_name", ""),
-                "asset_display_name": binding_payloads.get(name, {}).get("asset_display_name", ""),
-                "asset_fingerprint": binding_payloads.get(name, {}).get("asset_fingerprint", ""),
-                "display_name": str((catgirls_payload.get(name) or {}).get("档案名") or name),
-            }
-            for name in character_names
-        ],
-    }
-
-
-def _load_staged_json_file(staged_entries: dict[str, Path], relative_path: str, *, required: bool = False) -> Any:
-    staged_path = staged_entries.get(relative_path)
-    if staged_path is None:
-        if required:
-            raise ValueError(f"cloudsave import requires {relative_path}")
-        return None
-    with open(staged_path, "r", encoding="utf-8") as file_obj:
-        return json.load(file_obj)
-
-
-def _parse_binding_payloads(staged_entries: dict[str, Path]) -> dict[str, dict[str, Any]]:
-    binding_payloads: dict[str, dict[str, Any]] = {}
-    for relative_path, staged_path in staged_entries.items():
-        if not relative_path.startswith("bindings/") or not relative_path.endswith(".json"):
-            continue
-        binding_name = Path(relative_path).stem
-        with open(staged_path, "r", encoding="utf-8") as file_obj:
-            payload = json.load(file_obj)
-        if not isinstance(payload, dict):
-            raise ValueError(f"{relative_path} must contain a JSON object")
-        payload_name = str(payload.get("character_name") or "").strip()
-        if payload_name and payload_name != binding_name:
-            raise ValueError(f"{relative_path} character_name does not match filename")
-        binding_payloads[binding_name] = payload
-    return binding_payloads
-
-
-def _parse_catalog_character_names(payload: Any) -> set[str]:
-    if payload is None:
-        return set()
-    if not isinstance(payload, dict):
-        raise ValueError("catalog/catgirls_index.json must contain a JSON object")
-    names: set[str] = set()
-    for entry in payload.get("characters") or []:
-        if not isinstance(entry, dict):
-            raise ValueError("catalog/catgirls_index.json contains a non-object entry")
-        name = str(entry.get("character_name") or "").strip()
-        if not name:
-            raise ValueError("catalog/catgirls_index.json contains an empty character_name")
-        names.add(name)
-    return names
-
-
-def _build_catalog_current_character_payload(*, current_character_name: str, exported_at: str, sequence_number: int) -> dict[str, Any]:
-    return {
-        "current_character_name": current_character_name,
-        "last_known_name": current_character_name,
-        "applied_at_utc": exported_at,
-        "entry_sequence_number": sequence_number,
-    }
-
-
-def _iso_from_timestamp(timestamp: float | None) -> str:
-    if timestamp is None:
-        return ""
-    return datetime.fromtimestamp(timestamp, timezone.utc).isoformat().replace("+00:00", "Z")
-
-
-def _max_mtime_iso(paths: list[Path]) -> str:
-    latest_timestamp: float | None = None
-    for path in paths:
-        try:
-            timestamp = path.stat().st_mtime
-        except OSError:
-            continue
-        if latest_timestamp is None or timestamp > latest_timestamp:
-            latest_timestamp = timestamp
-    return _iso_from_timestamp(latest_timestamp)
-
-
-def _memory_file_hashes_from_root(root_dir: Path) -> dict[str, str]:
-    hashes: dict[str, str] = {}
-    for filename in MANAGED_MEMORY_FILENAMES:
-        path = root_dir / filename
-        if path.is_file():
-            hashes[filename] = _compute_managed_memory_file_hash(path)
-    return hashes
-
-
-def _compute_managed_memory_file_hash(path: Path) -> str:
-    if path.name != "time_indexed.db" or not _looks_like_sqlite_database(path):
-        return _sha256_file(path)
-
-    temp_root = Path(tempfile.mkdtemp(prefix="cloudsave-hash-"))
-    shadow_copy_path = temp_root / path.name
-    try:
-        _run_sqlite_shadow_copy(path, shadow_copy_path)
-        return _sha256_file(shadow_copy_path)
-    except sqlite3.DatabaseError as exc:
-        logger.warning(
-            "Falling back to direct SQLite file hash for %s after shadow-copy failure: %s",
-            path,
-            exc,
-        )
-        return _sha256_file(path)
-    finally:
-        shutil.rmtree(temp_root, ignore_errors=True)
-
-
-def _stable_binding_payload_for_fingerprint(binding_payload: Any) -> dict[str, Any]:
-    if not isinstance(binding_payload, dict):
-        return {}
-    stable_keys = (
-        "character_name",
-        "model_type",
-        "asset_source",
-        "asset_source_id",
-        "model_ref",
-        "asset_display_name",
-        "fallback_model_ref",
-        "experience_overrides",
-    )
-    return {
-        key: deepcopy(binding_payload.get(key))
-        for key in stable_keys
-        if key in binding_payload
-    }
-
-
-def _build_character_payload_fingerprint(
-    *,
-    character_name: str,
-    character_payload: Any,
-    binding_payload: Any,
-    memory_hashes: dict[str, str],
-) -> str:
-    fingerprint_payload = {
-        "schema_version": 1,
-        "character_name": str(character_name or ""),
-        "character_payload": deepcopy(character_payload) if isinstance(character_payload, dict) else {},
-        "binding_payload": _stable_binding_payload_for_fingerprint(binding_payload),
-        "memory_files": dict(sorted((memory_hashes or {}).items())),
-    }
-    return "sha256:" + _sha256_bytes(_json_canonical_dumps(fingerprint_payload).encode("utf-8"))
-
-
-def _build_character_summary_warnings(*, asset_state: str, warning_scope: str) -> list[str]:
-    warnings: list[str] = []
-    if asset_state in {"import_required", "downloadable", "missing"}:
-        if warning_scope == "local":
-            warnings.append("local_resource_missing_on_this_device")
-        elif warning_scope == "cloud":
-            warnings.append("cloud_resource_may_be_missing_after_download")
-    return warnings
-
-
-def _build_character_meta_payload(
-    *,
-    character_name: str,
-    binding_payload: dict[str, Any],
-    payload_fingerprint: str,
-    sequence_number: int,
-    exported_at: str,
-    client_id: str,
-    device_id: str,
-) -> dict[str, Any]:
-    return {
-        "schema_version": 1,
-        "character_name": character_name,
-        "payload_fingerprint": payload_fingerprint,
-        "updated_at_utc": exported_at,
-        "sequence_number": int(sequence_number),
-        "source_client_id": str(client_id or ""),
-        "source_device_id": str(device_id or ""),
-        "asset_state": str(binding_payload.get("asset_state") or ""),
-        "asset_source": str(binding_payload.get("asset_source") or ""),
-        "asset_source_id": str(binding_payload.get("asset_source_id") or ""),
-        "origin_source": str(binding_payload.get("origin_source") or ""),
-        "origin_source_id": str(binding_payload.get("origin_source_id") or ""),
-        "origin_model_ref": str(binding_payload.get("origin_model_ref") or ""),
-        "origin_display_name": str(binding_payload.get("origin_display_name") or ""),
-    }
-
-
-def _stage_single_character_cloudsave_entries(
-    config_manager,
-    stage_root: Path,
-    *,
-    character_name: str,
-    character_payload: dict[str, Any],
-    binding_payload: dict[str, Any],
-    sequence_number: int,
-    exported_at: str,
-    client_id: str,
-    device_id: str,
-) -> tuple[dict[str, Path], dict[str, Any]]:
-    staged_entries: dict[str, Path] = {}
-    object_root = f"characters/{character_name}"
-
-    memory_root = Path(config_manager.memory_dir) / character_name
-    memory_hashes: dict[str, str] = {}
-    for filename in MANAGED_MEMORY_FILENAMES:
-        source_path = memory_root / filename
-        if not source_path.is_file():
-            continue
-        relative_path = f"{object_root}/memory/{filename}"
-        staged_path = _stage_memory_file(stage_root, relative_path, source_path)
-        staged_entries[relative_path] = staged_path
-        memory_hashes[filename] = _sha256_file(staged_path)
-
-    payload_fingerprint = _build_character_payload_fingerprint(
-        character_name=character_name,
-        character_payload=character_payload,
-        binding_payload=binding_payload,
-        memory_hashes=memory_hashes,
-    )
-    meta_payload = _build_character_meta_payload(
-        character_name=character_name,
-        binding_payload=binding_payload,
-        payload_fingerprint=payload_fingerprint,
-        sequence_number=sequence_number,
-        exported_at=exported_at,
-        client_id=client_id,
-        device_id=device_id,
-    )
-
-    staged_entries[f"{object_root}/profile.json"] = _stage_json_file(
-        stage_root,
-        f"{object_root}/profile.json",
-        character_payload,
-    )
-    staged_entries[f"{object_root}/binding.json"] = _stage_json_file(
-        stage_root,
-        f"{object_root}/binding.json",
-        binding_payload,
-    )
-    staged_entries[f"{object_root}/meta.json"] = _stage_json_file(
-        stage_root,
-        f"{object_root}/meta.json",
-        meta_payload,
-    )
-    return staged_entries, meta_payload
-
-
-def _build_local_character_snapshot(
-    config_manager,
-    *,
-    character_name: str,
-    character_payload: dict[str, Any],
-    characters_config_path: Path,
-    workshop_origin_index: dict[str, list[dict[str, Any]]] | None = None,
-) -> dict[str, Any]:
-    binding_payload = _derive_character_binding_summary(
-        config_manager,
-        character_name,
-        character_payload,
-        workshop_origin_index=workshop_origin_index,
-    )
-    memory_root = Path(config_manager.memory_dir) / character_name
-    memory_hashes = _memory_file_hashes_from_root(memory_root)
-    updated_paths = [characters_config_path]
-    updated_paths.extend(memory_root / filename for filename in memory_hashes)
-    return {
-        "character_name": character_name,
-        "display_name": str(character_payload.get("档案名") or character_name),
-        "model_type": str(binding_payload.get("model_type") or ""),
-        "asset_source": str(binding_payload.get("asset_source") or ""),
-        "asset_source_id": str(binding_payload.get("asset_source_id") or ""),
-        "asset_state": str(binding_payload.get("asset_state") or ""),
-        "origin_source": str(binding_payload.get("origin_source") or ""),
-        "origin_source_id": str(binding_payload.get("origin_source_id") or ""),
-        "origin_model_ref": str(binding_payload.get("origin_model_ref") or ""),
-        "origin_display_name": str(binding_payload.get("origin_display_name") or ""),
-        "updated_at_utc": _max_mtime_iso(updated_paths),
-        "fingerprint": _build_character_payload_fingerprint(
-            character_name=character_name,
-            character_payload=character_payload,
-            binding_payload=binding_payload,
-            memory_hashes=memory_hashes,
-        ),
-        "warnings": _build_character_summary_warnings(
-            asset_state=str(binding_payload.get("asset_state") or ""),
-            warning_scope="local",
-        ),
-    }
-
-
-def _collect_cloudsave_catalog_entries(config_manager) -> dict[str, dict[str, Any]]:
-    payload = _load_json_if_exists(config_manager.cloudsave_catalog_dir / "catgirls_index.json")
-    if not isinstance(payload, dict):
-        return {}
-    entries: dict[str, dict[str, Any]] = {}
-    for entry in payload.get("characters") or []:
-        if not isinstance(entry, dict):
-            continue
-        character_name = str(entry.get("character_name") or "").strip()
-        if not character_name:
-            continue
-        entries[character_name] = entry
-    return entries
-
-
-def _load_cloudsave_tombstone_names(config_manager) -> set[str]:
-    tombstones_payload = _load_json_if_exists(config_manager.cloudsave_catalog_dir / "character_tombstones.json")
-    tombstones_state = _normalize_tombstones_state(tombstones_payload)
-    return {
-        entry["character_name"]
-        for entry in tombstones_state.get("tombstones") or []
-        if isinstance(entry, dict) and entry.get("character_name")
-    }
-
-
-def _load_cloudsave_sharded_character_unit(config_manager, character_name: str) -> dict[str, Any] | None:
-    object_dir = config_manager.cloudsave_dir / "characters" / character_name
-    profile_path = object_dir / "profile.json"
-    if not profile_path.is_file():
-        return None
-
-    profile_payload = _load_json_if_exists(profile_path)
-    if not isinstance(profile_payload, dict):
-        raise ValueError(f"cloudsave shard profile is invalid for {character_name}")
-
-    binding_payload = _load_json_if_exists(object_dir / "binding.json")
-    if binding_payload is not None and not isinstance(binding_payload, dict):
-        raise ValueError(f"cloudsave shard binding is invalid for {character_name}")
-
-    meta_payload = _load_json_if_exists(object_dir / "meta.json")
-    if meta_payload is not None and not isinstance(meta_payload, dict):
-        raise ValueError(f"cloudsave shard meta is invalid for {character_name}")
-
-    memory_files: dict[str, Path] = {}
-    memory_dir = object_dir / "memory"
-    if memory_dir.is_dir():
-        for filename in MANAGED_MEMORY_FILENAMES:
-            source_path = memory_dir / filename
-            if source_path.is_file():
-                memory_files[filename] = source_path
-
-    return {
-        "character_name": character_name,
-        "profile": profile_payload,
-        "binding": binding_payload or {},
-        "meta": meta_payload or {},
-        "memory_files": memory_files,
-    }
-
-
-def _collect_cloudsave_meta_payloads(config_manager) -> dict[str, dict[str, Any]]:
-    meta_payloads: dict[str, dict[str, Any]] = {}
-    sharded_root = config_manager.cloudsave_dir / "characters"
-    if not sharded_root.is_dir():
-        return meta_payloads
-    for child in sorted(sharded_root.iterdir()):
-        if not child.is_dir():
-            continue
-        payload = _load_json_if_exists(child / "meta.json")
-        if isinstance(payload, dict):
-            meta_payloads[child.name] = payload
-    return meta_payloads
-
-
-def _collect_cloudsave_binding_payloads(config_manager) -> dict[str, dict[str, Any]]:
-    binding_payloads: dict[str, dict[str, Any]] = {}
-    sharded_root = config_manager.cloudsave_dir / "characters"
-    if sharded_root.is_dir():
-        for child in sorted(sharded_root.iterdir()):
-            if not child.is_dir():
-                continue
-            payload = _load_json_if_exists(child / "binding.json")
-            if isinstance(payload, dict):
-                binding_payloads[child.name] = payload
-    bindings_dir = config_manager.cloudsave_bindings_dir
-    if not bindings_dir.is_dir():
-        bindings_dir = None
-    if bindings_dir is not None:
-        for path in sorted(bindings_dir.glob("*.json")):
-            payload = _load_json_if_exists(path)
-            if not isinstance(payload, dict):
-                continue
-            character_name = str(payload.get("character_name") or path.stem).strip()
-            if not character_name or character_name in binding_payloads:
-                continue
-            binding_payloads[character_name] = payload
-    return binding_payloads
-
-
-def _collect_cloudsave_memory_hashes(config_manager, character_name: str) -> tuple[dict[str, str], list[Path]]:
-    sharded_memory_root = config_manager.cloudsave_dir / "characters" / character_name / "memory"
-    sharded_hashes = _memory_file_hashes_from_root(sharded_memory_root)
-    if sharded_hashes:
-        return sharded_hashes, [sharded_memory_root / filename for filename in sharded_hashes]
-
-    legacy_memory_root = config_manager.cloudsave_memory_dir / character_name
-    legacy_hashes = _memory_file_hashes_from_root(legacy_memory_root)
-    return legacy_hashes, [legacy_memory_root / filename for filename in legacy_hashes]
-
-
-def _load_cloudsave_character_unit(config_manager, character_name: str) -> dict[str, Any] | None:
-    tombstone_names = _load_cloudsave_tombstone_names(config_manager)
-    if character_name in tombstone_names:
-        return None
-
-    sharded_unit = _load_cloudsave_sharded_character_unit(config_manager, character_name)
-    if sharded_unit is not None:
-        return sharded_unit
-
-    characters_payload = _load_json_if_exists(config_manager.cloudsave_profiles_dir / "characters.json")
-    if not isinstance(characters_payload, dict):
-        return None
-    character_payload = (characters_payload.get("猫娘") or {}).get(character_name)
-    if not isinstance(character_payload, dict):
-        return None
-
-    binding_payload = _load_json_if_exists(config_manager.cloudsave_bindings_dir / f"{character_name}.json")
-    if binding_payload is not None and not isinstance(binding_payload, dict):
-        raise ValueError(f"cloudsave binding payload is invalid for {character_name}")
-
-    memory_files: dict[str, Path] = {}
-    memory_dir = config_manager.cloudsave_memory_dir / character_name
-    for filename in MANAGED_MEMORY_FILENAMES:
-        source_path = memory_dir / filename
-        if source_path.is_file():
-            memory_files[filename] = source_path
-
-    detail = build_cloudsave_character_detail(config_manager, character_name)
-    return {
-        "character_name": character_name,
-        "profile": character_payload,
-        "binding": binding_payload or {},
-        "meta": {
-            "schema_version": 1,
-            "character_name": character_name,
-            "payload_fingerprint": str((((detail or {}).get("cloud_summary") or {}).get("fingerprint")) or ""),
-            "updated_at_utc": str((((detail or {}).get("cloud_summary") or {}).get("updated_at_utc")) or ""),
-            "sequence_number": 0,
-            "source_client_id": "",
-            "source_device_id": "",
-            "asset_state": str((((detail or {}).get("cloud_summary") or {}).get("asset_state")) or ""),
-            "asset_source": str((((detail or {}).get("cloud_summary") or {}).get("asset_source")) or ""),
-            "asset_source_id": str((((detail or {}).get("cloud_summary") or {}).get("asset_source_id")) or ""),
-            "origin_source": str((((detail or {}).get("cloud_summary") or {}).get("origin_source")) or ""),
-            "origin_source_id": str((((detail or {}).get("cloud_summary") or {}).get("origin_source_id")) or ""),
-            "origin_model_ref": str((((detail or {}).get("cloud_summary") or {}).get("origin_model_ref")) or ""),
-            "origin_display_name": str((((detail or {}).get("cloud_summary") or {}).get("origin_display_name")) or ""),
-        },
-        "memory_files": memory_files,
-    }
-
-
-def _build_cloud_character_snapshot(
-    config_manager,
-    *,
-    character_name: str,
-    character_payload: dict[str, Any],
-    binding_payloads: dict[str, dict[str, Any]],
-    meta_payloads: dict[str, dict[str, Any]],
-    manifest_exported_at: str,
-    catalog_entries: dict[str, dict[str, Any]],
-    workshop_origin_index: dict[str, list[dict[str, Any]]] | None = None,
-) -> dict[str, Any]:
-    binding_payload = deepcopy(binding_payloads.get(character_name) or {})
-    meta_payload = deepcopy(meta_payloads.get(character_name) or {})
-    memory_hashes, memory_paths = _collect_cloudsave_memory_hashes(config_manager, character_name)
-    object_dir = config_manager.cloudsave_dir / "characters" / character_name
-    payload_paths = [
-        object_dir / "profile.json",
-        object_dir / "binding.json",
-        object_dir / "meta.json",
-        config_manager.cloudsave_profiles_dir / "characters.json",
-        config_manager.cloudsave_bindings_dir / f"{character_name}.json",
-    ]
-    payload_paths.extend(memory_paths)
-    catalog_entry = catalog_entries.get(character_name) or {}
-    asset_state = str(
-        binding_payload.get("asset_state")
-        or catalog_entry.get("asset_state")
-        or meta_payload.get("asset_state")
-        or ""
-    )
-    default_origin_payload = _derive_character_origin_metadata(
-        config_manager,
-        character_name=character_name,
-        character_payload=character_payload,
-        model_type=str(binding_payload.get("model_type") or catalog_entry.get("model_type") or ""),
-        workshop_origin_index=workshop_origin_index,
-    )
-    updated_at_utc = str(
-        meta_payload.get("updated_at_utc")
-        or _max_mtime_iso(payload_paths)
-        or manifest_exported_at
-    )
-    return {
-        "character_name": character_name,
-        "display_name": str(character_payload.get("档案名") or catalog_entry.get("display_name") or character_name),
-        "model_type": str(
-            binding_payload.get("model_type")
-            or catalog_entry.get("model_type")
-            or ""
-        ),
-        "asset_source": str(
-            binding_payload.get("asset_source")
-            or catalog_entry.get("asset_source")
-            or meta_payload.get("asset_source")
-            or ""
-        ),
-        "asset_source_id": str(
-            binding_payload.get("asset_source_id")
-            or catalog_entry.get("asset_source_id")
-            or meta_payload.get("asset_source_id")
-            or ""
-        ),
-        "asset_state": asset_state,
-        "origin_source": str(
-            binding_payload.get("origin_source")
-            or catalog_entry.get("origin_source")
-            or meta_payload.get("origin_source")
-            or default_origin_payload.get("origin_source")
-            or ""
-        ),
-        "origin_source_id": str(
-            binding_payload.get("origin_source_id")
-            or catalog_entry.get("origin_source_id")
-            or meta_payload.get("origin_source_id")
-            or default_origin_payload.get("origin_source_id")
-            or ""
-        ),
-        "origin_model_ref": str(
-            binding_payload.get("origin_model_ref")
-            or catalog_entry.get("origin_model_ref")
-            or meta_payload.get("origin_model_ref")
-            or default_origin_payload.get("origin_model_ref")
-            or ""
-        ),
-        "origin_display_name": str(
-            binding_payload.get("origin_display_name")
-            or catalog_entry.get("origin_display_name")
-            or meta_payload.get("origin_display_name")
-            or default_origin_payload.get("origin_display_name")
-            or ""
-        ),
-        "updated_at_utc": updated_at_utc,
-        "fingerprint": _build_character_payload_fingerprint(
-            character_name=character_name,
-            character_payload=character_payload,
-            binding_payload=binding_payload,
-            memory_hashes=memory_hashes,
-        ),
-        "warnings": _build_character_summary_warnings(
-            asset_state=asset_state,
-            warning_scope="cloud",
-        ),
-    }
-
-
-def _load_cloudsave_character_payloads(config_manager) -> tuple[dict[str, dict[str, Any]], set[str]]:
-    tombstone_names = _load_cloudsave_tombstone_names(config_manager)
-    cloud_characters: dict[str, dict[str, Any]] = {}
-
-    characters_payload = _load_json_if_exists(config_manager.cloudsave_profiles_dir / "characters.json")
-    if isinstance(characters_payload, dict):
-        for character_name, character_payload in (characters_payload.get("猫娘") or {}).items():
-            if character_name in tombstone_names or not isinstance(character_payload, dict):
-                continue
-            cloud_characters[character_name] = character_payload
-
-    sharded_root = config_manager.cloudsave_dir / "characters"
-    if sharded_root.is_dir():
-        for child in sorted(sharded_root.iterdir()):
-            if not child.is_dir():
-                continue
-            payload = _load_json_if_exists(child / "profile.json")
-            if child.name in tombstone_names or not isinstance(payload, dict):
-                continue
-            # Prefer per-character shards when both formats are present.
-            cloud_characters[child.name] = payload
-
-    return cloud_characters, tombstone_names
-
-
-def _merge_character_summary_item(
-    *,
-    character_name: str,
-    local_summary: dict[str, Any] | None,
-    cloud_summary: dict[str, Any] | None,
-) -> dict[str, Any]:
-    local_exists = local_summary is not None
-    cloud_exists = cloud_summary is not None
-
-    local_fingerprint = str((local_summary or {}).get("fingerprint") or "")
-    cloud_fingerprint = str((cloud_summary or {}).get("fingerprint") or "")
-    relation_state = "local_only"
-    if local_exists and cloud_exists:
-        relation_state = "matched" if local_fingerprint and local_fingerprint == cloud_fingerprint else "diverged"
-    elif cloud_exists:
-        relation_state = "cloud_only"
-
-    available_actions: list[str] = []
-    if relation_state == "local_only":
-        available_actions = ["upload"]
-    elif relation_state == "cloud_only":
-        available_actions = ["download"]
-    elif relation_state == "diverged":
-        available_actions = ["upload", "download"]
-
-    # Warning text is phrased for the current device, so when a local character exists
-    # we should trust the local asset check and avoid leaking cloud-side warning state.
-    warnings_source = local_summary if local_exists else cloud_summary
-    warnings = list((warnings_source or {}).get("warnings") or [])
-
-    deduped_warnings: list[str] = []
-    for warning in warnings:
-        if warning not in deduped_warnings:
-            deduped_warnings.append(warning)
-
-    primary_summary = local_summary or cloud_summary or {}
-    return {
-        "character_name": character_name,
-        "display_name": str(primary_summary.get("display_name") or character_name),
-        "relation_state": relation_state,
-        "local_exists": local_exists,
-        "cloud_exists": cloud_exists,
-        "model_type": str(primary_summary.get("model_type") or ""),
-        "asset_source": str(primary_summary.get("asset_source") or ""),
-        "asset_source_id": str(primary_summary.get("asset_source_id") or ""),
-        "local_asset_source": str((local_summary or {}).get("asset_source") or ""),
-        "local_asset_source_id": str((local_summary or {}).get("asset_source_id") or ""),
-        "cloud_asset_source": str((cloud_summary or {}).get("asset_source") or ""),
-        "cloud_asset_source_id": str((cloud_summary or {}).get("asset_source_id") or ""),
-        "local_origin_source": str((local_summary or {}).get("origin_source") or ""),
-        "local_origin_source_id": str((local_summary or {}).get("origin_source_id") or ""),
-        "local_origin_display_name": str((local_summary or {}).get("origin_display_name") or ""),
-        "cloud_origin_source": str((cloud_summary or {}).get("origin_source") or ""),
-        "cloud_origin_source_id": str((cloud_summary or {}).get("origin_source_id") or ""),
-        "cloud_origin_display_name": str((cloud_summary or {}).get("origin_display_name") or ""),
-        "local_asset_state": str((local_summary or {}).get("asset_state") or ""),
-        "cloud_asset_state": str((cloud_summary or {}).get("asset_state") or ""),
-        "local_updated_at_utc": str((local_summary or {}).get("updated_at_utc") or ""),
-        "cloud_updated_at_utc": str((cloud_summary or {}).get("updated_at_utc") or ""),
-        "local_fingerprint": local_fingerprint,
-        "cloud_fingerprint": cloud_fingerprint,
-        "available_actions": available_actions,
-        "warnings": deduped_warnings,
-    }
-
-
-def _build_cloudsave_summary_state(
-    config_manager,
-) -> tuple[dict[str, Any], dict[str, dict[str, Any]], dict[str, dict[str, Any]]]:
-    if not is_cloudsave_disabled():
-        bootstrap_local_cloudsave_environment(config_manager)
-
-    characters_payload = config_manager.load_characters()
-    local_character_map = characters_payload.get("猫娘") or {}
-    current_character_name = str(characters_payload.get("当前猫娘") or "")
-    characters_config_path = Path(config_manager.get_runtime_config_path("characters.json"))
-    workshop_origin_index = _collect_workshop_character_origin_candidates(config_manager)
-    local_summaries = {
-        character_name: _build_local_character_snapshot(
-            config_manager,
-            character_name=character_name,
-            character_payload=character_payload,
-            characters_config_path=characters_config_path,
-            workshop_origin_index=workshop_origin_index,
-        )
-        for character_name, character_payload in sorted(local_character_map.items())
-        if isinstance(character_payload, dict)
-    }
-
-    provider_available = is_cloudsave_provider_available(config_manager)
-    cloud_summaries: dict[str, dict[str, Any]] = {}
-    if provider_available:
-        manifest = load_cloudsave_manifest(config_manager)
-        cloud_character_map, _tombstone_names = _load_cloudsave_character_payloads(config_manager)
-        catalog_entries = _collect_cloudsave_catalog_entries(config_manager)
-        binding_payloads = _collect_cloudsave_binding_payloads(config_manager)
-        meta_payloads = _collect_cloudsave_meta_payloads(config_manager)
-        cloud_summaries = {
-            character_name: _build_cloud_character_snapshot(
-                config_manager,
-                character_name=character_name,
-                character_payload=character_payload,
-                binding_payloads=binding_payloads,
-                meta_payloads=meta_payloads,
-                manifest_exported_at=str(manifest.get("exported_at_utc") or ""),
-                catalog_entries=catalog_entries,
-                workshop_origin_index=workshop_origin_index,
-            )
-            for character_name, character_payload in sorted(cloud_character_map.items())
-        }
-
-    all_names = sorted(set(local_summaries) | set(cloud_summaries))
-    items = [
-        _merge_character_summary_item(
-            character_name=character_name,
-            local_summary=local_summaries.get(character_name),
-            cloud_summary=cloud_summaries.get(character_name),
-        )
-        for character_name in all_names
-    ]
-    summary = {
-        "success": True,
-        "provider_available": provider_available,
-        "current_character_name": current_character_name,
-        "items": items,
-    }
-    return summary, local_summaries, cloud_summaries
-
-
-def build_cloudsave_summary(config_manager) -> dict[str, Any]:
-    summary, _local_summaries, _cloud_summaries = _build_cloudsave_summary_state(config_manager)
-    return summary
-
-
-def build_cloudsave_character_detail(config_manager, character_name: str) -> dict[str, Any] | None:
-    summary, local_summaries, cloud_summaries = _build_cloudsave_summary_state(config_manager)
-    for item in summary.get("items") or []:
-        if item.get("character_name") == character_name:
-            return {
-                "success": True,
-                "provider_available": bool(summary.get("provider_available", True)),
-                "current_character_name": str(summary.get("current_character_name") or ""),
-                "item": item,
-                "local_summary": deepcopy(local_summaries.get(character_name)),
-                "cloud_summary": deepcopy(cloud_summaries.get(character_name)),
-            }
-    return None
-
-
-def _assert_single_character_name_safe(character_name: str, *, context: str) -> None:
-    audit_result = audit_cloudsave_character_names([character_name])
-    try:
-        _raise_for_name_audit(audit_result, context=context)
-    except ValueError as exc:
-        raise CloudsaveOperationError(
-            "NAME_AUDIT_FAILED",
-            str(exc),
-            character_name=character_name,
-        ) from exc
-
-
-def export_cloudsave_character_unit(config_manager, character_name: str, *, overwrite: bool = False) -> dict[str, Any]:
-    if is_cloudsave_disabled():
-        _raise_cloudsave_disabled("single_character_upload", character_name=character_name)
-    bootstrap_local_cloudsave_environment(config_manager)
-    _assert_single_character_name_safe(character_name, context="single_character_upload")
-
-    with cloud_apply_fence(
-        config_manager,
-        mode=ROOT_MODE_BOOTSTRAP_IMPORTING,
-        reason=f"single_character_upload:{character_name}",
-    ):
-        characters_payload = config_manager.load_characters()
-        character_payload = (characters_payload.get("猫娘") or {}).get(character_name)
-        if not isinstance(character_payload, dict):
-            raise CloudsaveOperationError(
-                "LOCAL_CHARACTER_NOT_FOUND",
-                f"local character not found: {character_name}",
-                character_name=character_name,
-            )
-
-        existing_cloud_unit = _load_cloudsave_character_unit(config_manager, character_name)
-        if existing_cloud_unit is not None and not overwrite:
-            raise CloudsaveOperationError(
-                "CLOUD_CHARACTER_EXISTS",
-                f"cloud character already exists: {character_name}",
-                character_name=character_name,
-            )
-
-        stage_root = _create_staging_workspace(config_manager, "single-export")
-        cloud_state = config_manager.load_cloudsave_local_state()
-        sequence_number = max(1, int(cloud_state.get("next_sequence_number") or 1))
-        exported_at = _utc_now_iso()
-        manifest = ensure_cloudsave_manifest(config_manager)
-        workshop_origin_index = _collect_workshop_character_origin_candidates(config_manager)
-        binding_payload = _derive_character_binding_summary(
-            config_manager,
-            character_name,
-            character_payload,
-            workshop_origin_index=workshop_origin_index,
-        )
-        local_summary = _build_local_character_snapshot(
-            config_manager,
-            character_name=character_name,
-            character_payload=character_payload,
-            characters_config_path=Path(config_manager.get_runtime_config_path("characters.json")),
-            workshop_origin_index=workshop_origin_index,
-        )
-
-        staged_entries: dict[str, Path] = {}
-        existing_cloud_character_map, _tombstone_names = _load_cloudsave_character_payloads(config_manager)
-        cloud_profiles_payload = _load_json_if_exists(config_manager.cloudsave_profiles_dir / "characters.json")
-        if not isinstance(cloud_profiles_payload, dict):
-            cloud_profiles_payload = {}
-        cloud_profiles_payload = deepcopy(cloud_profiles_payload)
-        merged_cloud_character_map = deepcopy(existing_cloud_character_map)
-        merged_cloud_character_map[character_name] = deepcopy(character_payload)
-        cloud_profiles_payload["猫娘"] = {
-            name: deepcopy(payload)
-            for name, payload in sorted(merged_cloud_character_map.items())
-        }
-        staged_entries["profiles/characters.json"] = _stage_json_file(
-            stage_root,
-            "profiles/characters.json",
-            cloud_profiles_payload,
-        )
-
-        staged_entries[f"bindings/{character_name}.json"] = _stage_json_file(
-            stage_root,
-            f"bindings/{character_name}.json",
-            binding_payload,
-        )
-
-        character_memory_dir = Path(config_manager.memory_dir) / character_name
-        staged_memory_relative_paths: set[str] = set()
-        for filename in MANAGED_MEMORY_FILENAMES:
-            source_path = character_memory_dir / filename
-            if not source_path.is_file():
-                continue
-            relative_path = f"memory/{character_name}/{filename}"
-            staged_entries[relative_path] = _stage_memory_file(stage_root, relative_path, source_path)
-            staged_memory_relative_paths.add(relative_path)
-
-        single_character_entries, meta_payload = _stage_single_character_cloudsave_entries(
-            config_manager,
-            stage_root,
-            character_name=character_name,
-            character_payload=character_payload,
-            binding_payload=binding_payload,
-            sequence_number=sequence_number,
-            exported_at=exported_at,
-            client_id=str(cloud_state.get("client_id", "")),
-            device_id=str(manifest.get("device_id", "")),
-        )
-        staged_entries.update(single_character_entries)
-
-        merged_binding_payloads = _collect_cloudsave_binding_payloads(config_manager)
-        merged_binding_payloads[character_name] = deepcopy(binding_payload)
-        updated_catalog_payload = _build_catalog_index_payload(
-            character_names=sorted(merged_cloud_character_map),
-            characters_payload=cloud_profiles_payload,
-            binding_payloads=merged_binding_payloads,
-            sequence_number=sequence_number,
-            exported_at=exported_at,
-        )
-        staged_entries["catalog/catgirls_index.json"] = _stage_json_file(
-            stage_root,
-            "catalog/catgirls_index.json",
-            updated_catalog_payload,
-        )
-
-        updated_tombstones_payload = _remove_tombstone_from_catalog_payload(
-            _load_json_if_exists(config_manager.cloudsave_catalog_dir / "character_tombstones.json"),
-            character_name=character_name,
-            sequence_number=sequence_number,
-            exported_at=exported_at,
-        )
-        staged_entries["catalog/character_tombstones.json"] = _stage_json_file(
-            stage_root,
-            "catalog/character_tombstones.json",
-            updated_tombstones_payload,
-        )
-
-        upload_tag = exported_at.replace(":", "").replace(".", "")
-        backup_root = config_manager.cloudsave_backups_dir / f"character-upload-{upload_tag}" / character_name
-
-        existing_cloud_memory_root = config_manager.cloudsave_memory_dir / character_name
-        existing_cloud_character_root = config_manager.cloudsave_dir / "characters" / character_name
-        delete_targets: set[Path] = set()
-        for base_dir in (existing_cloud_memory_root, existing_cloud_character_root / "memory"):
-            if not base_dir.is_dir():
-                continue
-            for child in base_dir.iterdir():
-                if not child.is_file():
-                    continue
-                if base_dir == existing_cloud_memory_root:
-                    relative_path = f"memory/{character_name}/{child.name}"
-                else:
-                    relative_path = f"characters/{character_name}/memory/{child.name}"
-                if relative_path not in staged_entries:
-                    delete_targets.add(child)
-
-        mutation_targets = {
-            config_manager.cloudsave_profiles_dir / "characters.json",
-            config_manager.cloudsave_bindings_dir / f"{character_name}.json",
-            config_manager.cloudsave_catalog_dir / "catgirls_index.json",
-            config_manager.cloudsave_catalog_dir / "character_tombstones.json",
-            config_manager.cloudsave_dir / "characters" / character_name,
-            config_manager.cloudsave_memory_dir / character_name,
-            config_manager.cloudsave_manifest_path,
-            config_manager.cloudsave_local_state_path,
-        }
-        backup_records = _snapshot_existing_targets(
-            config_manager,
-            backup_root,
-            mutation_targets | delete_targets,
-        )
-
-        try:
-            for relative_path, staged_path in staged_entries.items():
-                _atomic_copy_file(staged_path, config_manager.cloudsave_dir / relative_path)
-
-            for target_path in sorted(delete_targets):
-                if target_path.exists():
-                    target_path.unlink()
-                    _cleanup_empty_parent_dirs(target_path, config_manager.cloudsave_dir)
-
-            manifest = _rebuild_cloudsave_manifest_from_disk(
-                config_manager,
-                sequence_number=sequence_number,
-                exported_at=exported_at,
-                client_id=str(cloud_state.get("client_id", "")),
-            )
-            cloud_state["next_sequence_number"] = sequence_number + 1
-            cloud_state["last_applied_manifest_fingerprint"] = str(manifest.get("fingerprint") or "")
-            cloud_state["last_successful_export_at"] = exported_at
-            config_manager.save_cloudsave_local_state(cloud_state)
-        except Exception:
-            _restore_backup_records(backup_records)
-            raise
-
-        detail = build_cloudsave_character_detail(config_manager, character_name)
-        return {
-            "character_name": character_name,
-            "sequence_number": sequence_number,
-            "meta": meta_payload,
-            "manifest": manifest,
-            "local_summary": local_summary,
-            "detail": detail,
-        }
-
-
-def import_cloudsave_character_unit(
-    config_manager,
-    character_name: str,
-    *,
-    overwrite: bool = False,
-    backup_before_overwrite: bool = True,
-) -> dict[str, Any]:
-    if is_cloudsave_disabled():
-        _raise_cloudsave_disabled("single_character_download", character_name=character_name)
-    bootstrap_local_cloudsave_environment(config_manager)
-    _assert_single_character_name_safe(character_name, context="single_character_download")
-
-    with cloud_apply_fence(
-        config_manager,
-        mode=ROOT_MODE_BOOTSTRAP_IMPORTING,
-        reason=f"single_character_download:{character_name}",
-    ):
-        cloud_unit = _load_cloudsave_character_unit(config_manager, character_name)
-        if cloud_unit is None:
-            raise CloudsaveOperationError(
-                "CLOUD_CHARACTER_NOT_FOUND",
-                f"cloud character not found: {character_name}",
-                character_name=character_name,
-            )
-
-        runtime_characters = config_manager.load_characters()
-        local_exists = character_name in (runtime_characters.get("猫娘") or {})
-        if local_exists and not overwrite:
-            raise CloudsaveOperationError(
-                "LOCAL_CHARACTER_EXISTS",
-                f"local character already exists: {character_name}",
-                character_name=character_name,
-            )
-
-        stage_root = _create_staging_workspace(config_manager, "single-import")
-        apply_time = _utc_now_iso()
-        updated_characters = deepcopy(runtime_characters)
-        updated_characters.setdefault("猫娘", {})
-        updated_characters["猫娘"][character_name] = deepcopy(cloud_unit["profile"])
-        current_character_name = str(updated_characters.get("当前猫娘") or "")
-        if not current_character_name:
-            updated_characters["当前猫娘"] = character_name
-        characters_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/profiles/characters.json",
-            updated_characters,
-        )
-
-        updated_tombstones_state = _remove_tombstone_from_state_payload(
-            config_manager.load_character_tombstones_state(),
-            character_name=character_name,
-        )
-        tombstones_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/state/character_tombstones.json",
-            updated_tombstones_state,
-        )
-
-        cloud_state = config_manager.load_cloudsave_local_state()
-        cloud_state["last_successful_import_at"] = apply_time
-        cloud_state_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/state/cloudsave_local_state.json",
-            cloud_state,
-        )
-
-        runtime_targets: dict[Path, Path] = {
-            Path(config_manager.get_runtime_config_path("characters.json")): characters_stage_path,
-            config_manager.character_tombstones_state_path: tombstones_stage_path,
-            config_manager.cloudsave_local_state_path: cloud_state_stage_path,
-        }
-        expected_memory_filenames: set[str] = set()
-        for filename, source_path in (cloud_unit.get("memory_files") or {}).items():
-            target_stage_path = _stage_file_copy(
-                stage_root,
-                f"__runtime__/memory/{character_name}/{filename}",
-                source_path,
-            )
-            runtime_targets[Path(config_manager.memory_dir) / character_name / filename] = target_stage_path
-            expected_memory_filenames.add(filename)
-
-        delete_file_targets: set[Path] = set()
-        target_memory_dir = Path(config_manager.memory_dir) / character_name
-        for filename in MANAGED_MEMORY_FILENAMES:
-            if filename in expected_memory_filenames:
-                continue
-            candidate = target_memory_dir / filename
-            if candidate.exists():
-                delete_file_targets.add(candidate)
-
-        backup_root = config_manager.cloudsave_backups_dir / f"character-download-{apply_time.replace(':', '').replace('.', '')}" / character_name
-        backup_targets = set(runtime_targets) | delete_file_targets
-        if backup_before_overwrite or not local_exists:
-            backup_targets.add(target_memory_dir)
-        backup_records = _snapshot_existing_targets(config_manager, backup_root, backup_targets)
-        _write_operation_backup_metadata(
-            config_manager,
-            backup_root,
-            operation="character_download",
-            character_name=character_name,
-            backup_records=backup_records,
-        )
-
-        try:
-            for target_path, staged_path in runtime_targets.items():
-                _apply_runtime_file(staged_path, target_path)
-
-            for target_path in sorted(delete_file_targets):
-                if target_path.exists():
-                    target_path.unlink()
-                    _cleanup_empty_parent_dirs(target_path, Path(config_manager.memory_dir))
-        except Exception:
-            _restore_backup_records(backup_records)
-            raise
-
-        detail = build_cloudsave_character_detail(config_manager, character_name)
-        return {
-            "character_name": character_name,
-            "applied_at_utc": apply_time,
-            "detail": detail,
-            "backup_path": str(backup_root),
-        }
-
-
-def _collect_memory_stage_entries(
-    config_manager,
-    stage_root: Path,
-    character_names: list[str],
-    *,
-    deadline_monotonic: float | None = None,
-    operation: str = "export",
-) -> dict[str, Path]:
-    staged_entries: dict[str, Path] = {}
-    for character_name in sorted(character_names):
-        character_dir = Path(config_manager.memory_dir) / character_name
-        for filename in MANAGED_MEMORY_FILENAMES:
-            _assert_deadline_not_exceeded(
-                deadline_monotonic,
-                operation=operation,
-                stage=f"stage_memory:{character_name}:{filename}",
-            )
-            source_path = character_dir / filename
-            if not source_path.is_file():
-                continue
-            relative_path = f"memory/{character_name}/{filename}"
-            staged_entries[relative_path] = _stage_memory_file(stage_root, relative_path, source_path)
-    return staged_entries
-
-
-def _managed_target_relative_path(config_manager, target_path: Path) -> Path:
-    normalized_target = Path(target_path).expanduser().resolve(strict=False)
-    runtime_root = Path(config_manager.app_docs_dir).expanduser().resolve(strict=False)
-    anchor_root = Path(getattr(config_manager, "anchor_root", config_manager.app_docs_dir)).expanduser().resolve(strict=False)
-
-    candidate_roots = [("runtime", runtime_root)]
-    if anchor_root != runtime_root:
-        candidate_roots.append(("anchor", anchor_root))
-    candidate_roots.sort(key=lambda item: len(item[1].parts), reverse=True)
-
-    for scope, root in candidate_roots:
-        try:
-            relative_path = normalized_target.relative_to(root)
-        except ValueError:
-            continue
-        return Path(scope) / relative_path
-
-    raise ValueError(f"unmanaged cloudsave backup target: {target_path}")
-
-
-def _resolve_managed_target_path(config_manager, relative_path: str) -> Path:
-    normalized_relative_path = str(relative_path or "").strip().replace("\\", "/")
-    if not normalized_relative_path:
-        raise ValueError("managed backup relative path is empty")
-
-    parts = Path(normalized_relative_path)
-    if not parts.parts or parts.is_absolute() or ".." in parts.parts:
-        raise ValueError("managed backup relative path is invalid")
-
-    scope = parts.parts[0]
-    suffix = Path(*parts.parts[1:]) if len(parts.parts) > 1 else Path()
-    if scope == "anchor":
-        root = Path(getattr(config_manager, "anchor_root", config_manager.app_docs_dir))
-    elif scope == "runtime":
-        root = Path(config_manager.app_docs_dir)
-    else:
-        # Backward compatibility for backups created before dual-root metadata was introduced.
-        root = Path(config_manager.app_docs_dir)
-        suffix = Path(normalized_relative_path)
-
-    resolved_root = root.expanduser().resolve(strict=False)
-    candidate = (root / suffix).expanduser().resolve(strict=False)
-    try:
-        candidate.relative_to(resolved_root)
-    except ValueError as exc:
-        raise ValueError("managed backup relative path escapes storage root") from exc
-    return candidate
-
-
-def _build_backup_path(config_manager, backup_root: Path, target_path: Path) -> Path:
-    return backup_root / _managed_target_relative_path(config_manager, target_path)
-
-
-def _snapshot_existing_targets(config_manager, backup_root: Path, targets: set[Path]) -> list[dict[str, Any]]:
-    backup_records: list[dict[str, Any]] = []
-    for target_path in sorted(targets, key=lambda path: (len(path.parts), str(path))):
-        relative_path = _managed_target_relative_path(config_manager, target_path)
-        record = {
-            "target": target_path,
-            "backup": None,
-            "is_dir": target_path.is_dir(),
-            "relative_path": str(relative_path).replace("\\", "/"),
-        }
-        if target_path.exists():
-            backup_path = _build_backup_path(config_manager, backup_root, target_path)
-            backup_path.parent.mkdir(parents=True, exist_ok=True)
-            if target_path.is_dir():
-                shutil.copytree(target_path, backup_path, dirs_exist_ok=True)
-            else:
-                shutil.copy2(target_path, backup_path)
-            record["backup"] = backup_path
-        backup_records.append(record)
-    return backup_records
-
-
-def _restore_backup_records(backup_records: list[dict[str, Any]]) -> None:
-    for record in sorted(backup_records, key=lambda item: len(item["target"].parts), reverse=True):
-        target_path = record["target"]
-        if target_path.exists():
-            if target_path.is_dir():
-                shutil.rmtree(target_path, ignore_errors=True)
-            else:
-                target_path.unlink()
-        backup_path = record.get("backup")
-        if backup_path is None or not backup_path.exists():
-            continue
-        if record.get("is_dir"):
-            shutil.copytree(backup_path, target_path, dirs_exist_ok=True)
-        else:
-            _apply_runtime_file(backup_path, target_path)
-
-
-def _write_operation_backup_metadata(
-    config_manager,
-    backup_root: Path,
-    *,
-    operation: str,
-    character_name: str,
-    backup_records: list[dict[str, Any]],
-) -> Path:
-    payload = {
-        "schema_version": 1,
-        "operation": operation,
-        "character_name": character_name,
-        "targets": [
-            {
-                "relative_path": str(record.get("relative_path") or ""),
-                "had_backup": record.get("backup") is not None,
-                "is_dir": bool(record.get("is_dir", False)),
-            }
-            for record in backup_records
-        ],
-    }
-    metadata_path = backup_root / "_operation.json"
-    atomic_write_json(metadata_path, payload, ensure_ascii=False, indent=2)
-    return metadata_path
-
-
-def restore_cloudsave_operation_backup(config_manager, backup_root: str | Path) -> None:
-    backup_root_path = Path(backup_root)
-    metadata = _load_json_if_exists(backup_root_path / "_operation.json")
-    if not isinstance(metadata, dict):
-        raise FileNotFoundError(f"cloudsave backup metadata missing: {backup_root_path}")
-
-    backup_records: list[dict[str, Any]] = []
-    for target in metadata.get("targets") or []:
-        if not isinstance(target, dict):
-            continue
-        relative_path = str(target.get("relative_path") or "").strip().replace("\\", "/")
-        if not relative_path:
-            continue
-        runtime_target = _resolve_managed_target_path(config_manager, relative_path)
-        backup_path = backup_root_path / relative_path
-        backup_records.append(
-            {
-                "target": runtime_target,
-                "backup": backup_path if bool(target.get("had_backup")) and backup_path.exists() else None,
-                "is_dir": bool(target.get("is_dir", False)),
-            }
-        )
-    _restore_backup_records(backup_records)
-
-
-def _rebuild_cloudsave_manifest_from_disk(
-    config_manager,
-    *,
-    sequence_number: int,
-    exported_at: str,
-    client_id: str,
-) -> dict[str, Any]:
-    manifest = ensure_cloudsave_manifest(config_manager)
-    files = {
-        relative_path: {
-            "sha256": _sha256_file(config_manager.cloudsave_dir / relative_path),
-            "size": (config_manager.cloudsave_dir / relative_path).stat().st_size,
-        }
-        for relative_path in sorted(_list_existing_cloudsave_files(config_manager))
-    }
-    manifest.update(
-        {
-            "schema_version": 1,
-            "min_reader_schema_version": 1,
-            "min_app_version": "",
-            "client_id": str(client_id or manifest.get("client_id", "")),
-            "device_id": str(manifest.get("device_id", "")),
-            "sequence_number": int(sequence_number),
-            "exported_at_utc": exported_at,
-            "files": files,
-        }
-    )
-    manifest["fingerprint"] = _build_manifest_fingerprint(
-        client_id=str(manifest.get("client_id", "")),
-        sequence_number=int(manifest.get("sequence_number") or 0),
-        files=files,
-    )
-    save_cloudsave_manifest(config_manager, manifest)
-    return manifest
-
-
-def _default_catalog_index_payload() -> dict[str, Any]:
-    return {
-        "schema_version": 1,
-        "sequence_number": 0,
-        "exported_at_utc": "",
-        "characters": [],
-    }
-
-
-def _default_tombstones_catalog_payload() -> dict[str, Any]:
-    return {
-        "schema_version": 1,
-        "sequence_number": 0,
-        "exported_at_utc": "",
-        "tombstones": [],
-    }
-
-
-def _upsert_catalog_character_entry(
-    catalog_payload: Any,
-    *,
-    character_entry: dict[str, Any],
-    sequence_number: int,
-    exported_at: str,
-) -> dict[str, Any]:
-    payload = deepcopy(catalog_payload) if isinstance(catalog_payload, dict) else _default_catalog_index_payload()
-    entries_by_name: dict[str, dict[str, Any]] = {}
-    for entry in payload.get("characters") or []:
-        if not isinstance(entry, dict):
-            continue
-        existing_name = str(entry.get("character_name") or "").strip()
-        if existing_name:
-            entries_by_name[existing_name] = deepcopy(entry)
-    entry_name = str(character_entry.get("character_name") or "").strip()
-    if entry_name:
-        entries_by_name[entry_name] = deepcopy(character_entry)
-    payload["schema_version"] = 1
-    payload["sequence_number"] = int(sequence_number)
-    payload["exported_at_utc"] = exported_at
-    payload["characters"] = [entries_by_name[name] for name in sorted(entries_by_name)]
-    return payload
-
-
-def _remove_tombstone_from_catalog_payload(
-    tombstones_payload: Any,
-    *,
-    character_name: str,
-    sequence_number: int,
-    exported_at: str,
-) -> dict[str, Any]:
-    payload = deepcopy(tombstones_payload) if isinstance(tombstones_payload, dict) else _default_tombstones_catalog_payload()
-    tombstones_state = _normalize_tombstones_state(payload)
-    filtered_tombstones = [
-        entry
-        for entry in tombstones_state.get("tombstones") or []
-        if str(entry.get("character_name") or "") != character_name
-    ]
-    return {
-        "schema_version": 1,
-        "sequence_number": int(sequence_number),
-        "exported_at_utc": exported_at,
-        "tombstones": filtered_tombstones,
-    }
-
-
-def _remove_tombstone_from_state_payload(
-    tombstones_payload: Any,
-    *,
-    character_name: str,
-) -> dict[str, Any]:
-    tombstones_state = _normalize_tombstones_state(tombstones_payload)
-    return {
-        "version": 1,
-        "tombstones": [
-            entry
-            for entry in tombstones_state.get("tombstones") or []
-            if str(entry.get("character_name") or "") != character_name
-        ],
-    }
-
-
-def export_local_cloudsave_snapshot(
-    config_manager,
-    *,
-    deadline_monotonic: float | None = None,
-) -> dict[str, Any]:
-    """Export the current local runtime truth into cloudsave/ with manifest-last semantics."""
-    if is_cloudsave_disabled():
-        _raise_cloudsave_disabled("local_cloudsave_export")
-    bootstrap_local_cloudsave_environment(config_manager)
-
-    with cloud_apply_fence(
-        config_manager,
-        mode=ROOT_MODE_BOOTSTRAP_IMPORTING,
-        reason="local_cloudsave_export",
-    ):
-        _assert_deadline_not_exceeded(
-            deadline_monotonic,
-            operation="export",
-            stage="prepare_export",
-        )
-        stage_root = _create_staging_workspace(config_manager, "export")
-        cloud_state = config_manager.load_cloudsave_local_state()
-        sequence_number = max(1, int(cloud_state.get("next_sequence_number") or 1))
-        exported_at = _utc_now_iso()
-
-        characters_payload = config_manager.load_characters()
-        conversation_settings = _extract_conversation_settings(config_manager)
-        tombstones_state = _load_local_tombstones_state(config_manager)
-        tombstones = tombstones_state.get("tombstones") or []
-        live_character_names = sorted((characters_payload.get("猫娘") or {}).keys())
-        live_name_set = set(live_character_names)
-        filtered_tombstones = [
-            tombstone
-            for tombstone in tombstones
-            if tombstone.get("character_name") not in live_name_set
-        ]
-        if filtered_tombstones != tombstones:
-            tombstones_state["tombstones"] = filtered_tombstones
-            tombstones_state = _save_local_tombstones_state(config_manager, tombstones_state)
-            tombstones = tombstones_state.get("tombstones") or []
-        tombstone_names = [tombstone["character_name"] for tombstone in tombstones]
-        name_audit = audit_cloudsave_character_names(live_character_names, tombstone_names)
-        _raise_for_name_audit(name_audit, context="export")
-        character_names = live_character_names
-        current_character_name = str(characters_payload.get("当前猫娘") or "")
-        workshop_origin_index = _collect_workshop_character_origin_candidates(config_manager)
-        binding_payloads = {
-            name: _derive_character_binding_summary(
-                config_manager,
-                name,
-                (characters_payload.get("猫娘") or {}).get(name, {}),
-                workshop_origin_index=workshop_origin_index,
-            )
-            for name in character_names
-        }
-
-        sensitive_findings = scan_for_sensitive_values(characters_payload, path="profiles.characters")
-        if sensitive_findings:
-            raise ValueError(f"sensitive values detected in export payload: {', '.join(sensitive_findings)}")
-
-        staged_entries: dict[str, Path] = {
-            "profiles/characters.json": _stage_json_file(stage_root, "profiles/characters.json", characters_payload),
-            "profiles/conversation_settings.json": _stage_json_file(
-                stage_root,
-                "profiles/conversation_settings.json",
-                conversation_settings,
-            ),
-            "catalog/catgirls_index.json": _stage_json_file(
-                stage_root,
-                "catalog/catgirls_index.json",
-                _build_catalog_index_payload(
-                    character_names=character_names,
-                    characters_payload=characters_payload,
-                    binding_payloads=binding_payloads,
-                    sequence_number=sequence_number,
-                    exported_at=exported_at,
-                ),
-            ),
-            "catalog/current_character.json": _stage_json_file(
-                stage_root,
-                "catalog/current_character.json",
-                _build_catalog_current_character_payload(
-                    current_character_name=current_character_name,
-                    exported_at=exported_at,
-                    sequence_number=sequence_number,
-                ),
-            ),
-            "catalog/character_tombstones.json": _stage_json_file(
-                stage_root,
-                "catalog/character_tombstones.json",
-                _make_tombstones_catalog_payload(
-                    tombstones=tombstones,
-                    sequence_number=sequence_number,
-                    exported_at=exported_at,
-                ),
-            ),
-        }
-        manifest = ensure_cloudsave_manifest(config_manager)
-        manifest_device_id = str(manifest.get("device_id", ""))
-        for name, binding_payload in binding_payloads.items():
-            _assert_deadline_not_exceeded(
-                deadline_monotonic,
-                operation="export",
-                stage=f"stage_character:{name}",
-            )
-            staged_entries[f"bindings/{name}.json"] = _stage_json_file(
-                stage_root,
-                f"bindings/{name}.json",
-                binding_payload,
-            )
-            single_character_entries, _meta_payload = _stage_single_character_cloudsave_entries(
-                config_manager,
-                stage_root,
-                character_name=name,
-                character_payload=(characters_payload.get("猫娘") or {}).get(name, {}),
-                binding_payload=binding_payload,
-                sequence_number=sequence_number,
-                exported_at=exported_at,
-                client_id=str(cloud_state.get("client_id", "")),
-                device_id=manifest_device_id,
-            )
-            staged_entries.update(single_character_entries)
-        staged_entries.update(
-            _collect_memory_stage_entries(
-                config_manager,
-                stage_root,
-                character_names,
-                deadline_monotonic=deadline_monotonic,
-                operation="export",
-            )
-        )
-
-        _assert_deadline_not_exceeded(
-            deadline_monotonic,
-            operation="export",
-            stage="finalize_manifest",
-        )
-        files = {
-            relative_path: {
-                "sha256": _sha256_file(staged_path),
-                "size": staged_path.stat().st_size,
-            }
-            for relative_path, staged_path in sorted(staged_entries.items())
-        }
-
-        manifest.update(
-            {
-                "schema_version": 1,
-                "min_reader_schema_version": 1,
-                "min_app_version": "",
-                "client_id": str(cloud_state.get("client_id", "")),
-                "device_id": str(manifest.get("device_id", "")),
-                "sequence_number": sequence_number,
-                "exported_at_utc": exported_at,
-                "files": files,
-            }
-        )
-        manifest["fingerprint"] = _build_manifest_fingerprint(
-            client_id=manifest["client_id"],
-            sequence_number=sequence_number,
-            files=files,
-        )
-
-        _assert_deadline_not_exceeded(
-            deadline_monotonic,
-            operation="export",
-            stage="apply_snapshot",
-        )
-        for relative_path, staged_path in staged_entries.items():
-            _atomic_copy_file(staged_path, config_manager.cloudsave_dir / relative_path)
-
-        stale_files = _list_existing_cloudsave_files(config_manager) - set(staged_entries)
-        for relative_path in sorted(stale_files):
-            target_path = config_manager.cloudsave_dir / relative_path
-            if target_path.exists():
-                target_path.unlink()
-                _cleanup_empty_parent_dirs(target_path, config_manager.cloudsave_dir)
-
-        save_cloudsave_manifest(config_manager, manifest)
-
-        cloud_state["next_sequence_number"] = sequence_number + 1
-        cloud_state["last_applied_manifest_fingerprint"] = manifest["fingerprint"]
-        cloud_state["last_successful_export_at"] = exported_at
-        config_manager.save_cloudsave_local_state(cloud_state)
-
-        return {
-            "manifest": manifest,
-            "staged_file_count": len(staged_entries),
-            "name_audit": name_audit,
-        }
-
-
-def import_local_cloudsave_snapshot(
-    config_manager,
-    *,
-    deadline_monotonic: float | None = None,
-    use_cloud_apply_fence: bool = True,
-) -> dict[str, Any]:
-    """Import the current local cloudsave snapshot back into runtime truth with rollback."""
-    if is_cloudsave_disabled():
-        _raise_cloudsave_disabled("local_cloudsave_import")
-    bootstrap_local_cloudsave_environment(config_manager)
-    fence_scope = (
-        cloud_apply_fence(
-            config_manager,
-            mode=ROOT_MODE_BOOTSTRAP_IMPORTING,
-            reason="local_cloudsave_import",
-        )
-        if use_cloud_apply_fence
-        else nullcontext()
-    )
-    with fence_scope:
-        _assert_deadline_not_exceeded(
-            deadline_monotonic,
-            operation="import",
-            stage="prepare_import",
-        )
-        manifest = load_cloudsave_manifest(config_manager)
-        manifest_files = manifest.get("files") or {}
-        if not isinstance(manifest_files, dict) or not manifest_files:
-            raise ValueError("cloudsave manifest does not contain any staged files")
-
-        stage_root = _create_staging_workspace(config_manager, "import")
-        staged_entries: dict[str, Path] = {}
-        for relative_path in sorted(manifest_files):
-            _assert_deadline_not_exceeded(
-                deadline_monotonic,
-                operation="import",
-                stage=f"stage_file:{relative_path}",
-            )
-            source_path = config_manager.cloudsave_dir / relative_path
-            if not source_path.is_file():
-                raise FileNotFoundError(f"cloudsave file missing from manifest: {relative_path}")
-            staged_entries[relative_path] = _stage_file_copy(stage_root, relative_path, source_path)
-
-        computed_files = {
-            relative_path: {
-                "sha256": _sha256_file(staged_path),
-                "size": staged_path.stat().st_size,
-            }
-            for relative_path, staged_path in sorted(staged_entries.items())
-        }
-        computed_fingerprint = _build_manifest_fingerprint(
-            client_id=str(manifest.get("client_id", "")),
-            sequence_number=int(manifest.get("sequence_number") or 0),
-            files=computed_files,
-        )
-        if manifest.get("fingerprint") and manifest["fingerprint"] != computed_fingerprint:
-            raise ValueError("cloudsave manifest fingerprint mismatch")
-
-        characters_payload = _load_staged_json_file(staged_entries, "profiles/characters.json", required=True)
-        if not isinstance(characters_payload, dict):
-            raise ValueError("profiles/characters.json must contain a JSON object")
-
-        conversation_settings = _load_staged_json_file(staged_entries, "profiles/conversation_settings.json") or {}
-        if not isinstance(conversation_settings, dict):
-            raise ValueError("profiles/conversation_settings.json must contain a JSON object")
-
-        binding_payloads = _parse_binding_payloads(staged_entries)
-        catalog_index_payload = _load_staged_json_file(staged_entries, "catalog/catgirls_index.json")
-        current_character_catalog_payload = _load_staged_json_file(staged_entries, "catalog/current_character.json")
-        tombstones_catalog_payload = _load_staged_json_file(staged_entries, "catalog/character_tombstones.json") or {}
-        tombstones_state = _normalize_tombstones_state(tombstones_catalog_payload)
-        tombstones = tombstones_state.get("tombstones") or []
-        tombstone_names = [tombstone["character_name"] for tombstone in tombstones]
-
-        sensitive_findings = scan_for_sensitive_values(characters_payload, path="profiles.characters")
-        if sensitive_findings:
-            raise ValueError(f"sensitive values detected in import payload: {', '.join(sensitive_findings)}")
-
-        character_map = deepcopy(characters_payload.get("猫娘") or {})
-        live_character_names = sorted(character_map.keys())
-        name_audit = audit_cloudsave_character_names(live_character_names, tombstone_names)
-        _raise_for_name_audit(name_audit, context="import")
-
-        catalog_character_names = _parse_catalog_character_names(catalog_index_payload)
-        if catalog_character_names and catalog_character_names != set(live_character_names):
-            raise ValueError("catalog/catgirls_index.json is inconsistent with profiles/characters.json")
-        if binding_payloads and set(binding_payloads) != set(live_character_names):
-            raise ValueError("bindings/ payloads are inconsistent with profiles/characters.json")
-
-        for tombstone_name in tombstone_names:
-            character_map.pop(tombstone_name, None)
-        characters_payload["猫娘"] = character_map
-
-        requested_current_name = str(characters_payload.get("当前猫娘") or "").strip()
-        if isinstance(current_character_catalog_payload, dict):
-            catalog_current_name = str(current_character_catalog_payload.get("current_character_name") or "").strip()
-            if catalog_current_name:
-                requested_current_name = catalog_current_name
-
-        imported_character_names = sorted(character_map.keys())
-        if requested_current_name and requested_current_name in character_map:
-            characters_payload["当前猫娘"] = requested_current_name
-        elif imported_character_names:
-            characters_payload["当前猫娘"] = imported_character_names[0]
-        else:
-            characters_payload["当前猫娘"] = ""
-        apply_time = _utc_now_iso()
-        backup_root = config_manager.cloudsave_backups_dir / f"import-{apply_time.replace(':', '').replace('.', '')}"
-
-        characters_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/profiles/characters.json",
-            characters_payload,
-        )
-        runtime_targets: dict[Path, Path] = {
-            Path(config_manager.get_runtime_config_path("characters.json")): characters_stage_path,
-        }
-
-        preferences_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/user_preferences.json",
-            _build_runtime_preferences_payload(config_manager, conversation_settings),
-        )
-        runtime_targets[Path(config_manager.get_runtime_config_path("user_preferences.json"))] = preferences_stage_path
-
-        for relative_path, staged_path in staged_entries.items():
-            if not relative_path.startswith("memory/"):
-                continue
-            parts = Path(relative_path).parts
-            if len(parts) != 3:
-                raise ValueError(f"unsupported cloudsave memory path: {relative_path}")
-            _, character_name, filename = parts
-            if character_name in tombstone_names:
-                continue
-            runtime_targets[Path(config_manager.memory_dir) / character_name / filename] = staged_path
-
-        cloud_state = config_manager.load_cloudsave_local_state()
-        cloud_state["last_applied_manifest_fingerprint"] = computed_fingerprint
-        cloud_state["last_successful_import_at"] = apply_time
-        cloud_state_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/state/cloudsave_local_state.json",
-            cloud_state,
-        )
-        runtime_targets[config_manager.cloudsave_local_state_path] = cloud_state_stage_path
-        tombstones_state_stage_path = _stage_json_file(
-            stage_root,
-            "__runtime__/state/character_tombstones.json",
-            tombstones_state,
-        )
-        runtime_targets[config_manager.character_tombstones_state_path] = tombstones_state_stage_path
-
-        delete_file_targets: set[Path] = set()
-        delete_dir_targets: set[Path] = set()
-        for character_name in imported_character_names:
-            character_dir = Path(config_manager.memory_dir) / character_name
-            for filename in MANAGED_MEMORY_FILENAMES:
-                relative_path = f"memory/{character_name}/{filename}"
-                target_path = character_dir / filename
-                if relative_path not in staged_entries and target_path.exists():
-                    delete_file_targets.add(target_path)
-
-        memory_root = Path(config_manager.memory_dir)
-        if memory_root.exists():
-            for child in memory_root.iterdir():
-                if child.is_dir() and child.name not in imported_character_names:
-                    delete_dir_targets.add(child)
-
-        backup_records: list[dict[str, Any]] = []
-        for target_path in sorted(
-            set(runtime_targets) | delete_file_targets | delete_dir_targets,
-            key=lambda path: len(path.parts),
-        ):
-            record = {
-                "target": target_path,
-                "backup": None,
-                "is_dir": target_path.is_dir(),
-            }
-            if target_path.exists():
-                backup_path = _build_backup_path(config_manager, backup_root, target_path)
-                backup_path.parent.mkdir(parents=True, exist_ok=True)
-                if target_path.is_dir():
-                    shutil.copytree(target_path, backup_path, dirs_exist_ok=True)
-                else:
-                    shutil.copy2(target_path, backup_path)
-                record["backup"] = backup_path
-            backup_records.append(record)
-
-        _assert_deadline_not_exceeded(
-            deadline_monotonic,
-            operation="import",
-            stage="apply_runtime",
-        )
-        try:
-            for target_path, staged_path in runtime_targets.items():
-                _apply_runtime_file(staged_path, target_path)
-
-            for target_path in sorted(delete_file_targets):
-                if target_path.exists():
-                    target_path.unlink()
-                    _cleanup_empty_parent_dirs(target_path, Path(config_manager.memory_dir))
-
-            for target_path in sorted(delete_dir_targets, key=lambda path: len(path.parts), reverse=True):
-                if target_path.exists():
-                    shutil.rmtree(target_path)
-
-            return {
-                "manifest_fingerprint": computed_fingerprint,
-                "applied_character_count": len(imported_character_names),
-                "name_audit": name_audit,
-            }
-        except Exception:
-            for record in sorted(backup_records, key=lambda item: len(item["target"].parts), reverse=True):
-                target_path = record["target"]
-                if target_path.exists():
-                    if target_path.is_dir():
-                        shutil.rmtree(target_path, ignore_errors=True)
-                    else:
-                        target_path.unlink()
-                backup_path = record["backup"]
-                if backup_path is None or not backup_path.exists():
-                    continue
-                if record["is_dir"]:
-                    shutil.copytree(backup_path, target_path, dirs_exist_ok=True)
-                else:
-                    _apply_runtime_file(backup_path, target_path)
-            raise
-
-
-def load_cloudsave_manifest(config_manager, default_value: dict[str, Any] | None = None) -> dict[str, Any]:
-    if default_value is None:
-        cloud_state = config_manager.load_cloudsave_local_state()
-        default_value = build_default_cloudsave_manifest(client_id=cloud_state.get("client_id", ""))
-    return config_manager._load_json_file(config_manager.cloudsave_manifest_path, default_value)
-
-
-def save_cloudsave_manifest(config_manager, data: dict[str, Any]) -> None:
-    config_manager.ensure_cloudsave_structure()
-    atomic_write_json(config_manager.cloudsave_manifest_path, data, ensure_ascii=False, indent=2)
-
-
-def _ensure_local_state_directory_or_raise(config_manager, context: str) -> None:
-    if config_manager.ensure_local_state_directory():
-        return
-    if hasattr(config_manager, "_raise_local_state_directory_error"):
-        config_manager._raise_local_state_directory_error(context)
-    diagnostic = getattr(config_manager, "_last_local_state_directory_error", None)
-    if diagnostic is not None:
-        raise diagnostic
-    raise OSError("failed to ensure local state directory")
-
-
-def ensure_cloudsave_manifest(config_manager, *, preserve_existing_client_id: bool = False) -> dict[str, Any]:
-    config_manager.ensure_cloudsave_structure()
-    cloud_state = config_manager.load_cloudsave_local_state()
-    manifest = load_cloudsave_manifest(
-        config_manager,
-        default_value=build_default_cloudsave_manifest(client_id=cloud_state.get("client_id", "")),
-    )
-    changed = False
-    current_client_id = str(manifest.get("client_id") or "")
-    expected_client_id = str(cloud_state.get("client_id", "") or "")
-    if not current_client_id:
-        manifest["client_id"] = expected_client_id
-        changed = True
-    elif not preserve_existing_client_id and current_client_id != expected_client_id:
-        manifest["client_id"] = cloud_state.get("client_id", "")
-        changed = True
-    if "schema_version" not in manifest:
-        manifest["schema_version"] = 1
-        changed = True
-    if "min_reader_schema_version" not in manifest:
-        manifest["min_reader_schema_version"] = 1
-        changed = True
-    if "min_app_version" not in manifest:
-        manifest["min_app_version"] = ""
-        changed = True
-    if "device_id" not in manifest:
-        manifest["device_id"] = ""
-        changed = True
-    if "sequence_number" not in manifest:
-        manifest["sequence_number"] = 0
-        changed = True
-    if "exported_at_utc" not in manifest:
-        manifest["exported_at_utc"] = ""
-        changed = True
-    if "files" not in manifest or not isinstance(manifest.get("files"), dict):
-        manifest["files"] = {}
-        changed = True
-    if "fingerprint" not in manifest:
-        manifest["fingerprint"] = ""
-        changed = True
-    if changed or not config_manager.cloudsave_manifest_path.exists():
-        save_cloudsave_manifest(config_manager, manifest)
-    return manifest
-
-
-def bootstrap_local_cloudsave_environment(config_manager) -> dict[str, Any]:
-    """Initialize phase-0 local cloudsave skeleton and state files."""
-    if is_cloudsave_disabled():
-        return {
-            "disabled": True,
-            "disabled_reason": cloudsave_disabled_reason(),
-            "root_state": config_manager.build_default_root_state(),
-            "cloudsave_local_state": config_manager.build_default_cloudsave_local_state(client_id=""),
-            "manifest": build_default_cloudsave_manifest(client_id=""),
-            "legacy_import": {
-                "migrated": False,
-                "source": "",
-                "copied_paths": [],
-                "backup_path": "",
-                "repair_reason": "",
-                "result": "cloudsave_disabled",
-            },
-        }
-
-    _ensure_local_state_directory_or_raise(config_manager, "preparing local cloudsave state")
-
-    if not config_manager.ensure_cloudsave_structure():
-        raise OSError("failed to ensure cloudsave directory structure")
-
-    config_manager.ensure_cloudsave_state_files()
-
-    root_state = config_manager.load_root_state()
-    if str(root_state.get("mode") or ROOT_MODE_NORMAL) == ROOT_MODE_DEFERRED_INIT:
-        cloud_state = config_manager.load_cloudsave_local_state()
-        cloud_changed = False
-        if not cloud_state.get("client_id"):
-            cloud_state["client_id"] = config_manager.build_default_cloudsave_local_state()["client_id"]
-            cloud_changed = True
-        next_seq = int(cloud_state.get("next_sequence_number") or 0)
-        if next_seq < 1:
-            cloud_state["next_sequence_number"] = 1
-            cloud_changed = True
-        if cloud_changed:
-            config_manager.save_cloudsave_local_state(cloud_state)
-
-        manifest = ensure_cloudsave_manifest(config_manager, preserve_existing_client_id=True)
-        return {
-            "root_state": root_state,
-            "cloudsave_local_state": config_manager.load_cloudsave_local_state(),
-            "manifest": manifest,
-            "legacy_import": {
-                "migrated": False,
-                "source": "",
-                "copied_paths": [],
-                "backup_path": "",
-                "repair_reason": "",
-                "result": "recovery_required",
-            },
-        }
-
-    legacy_import = import_legacy_runtime_root_if_needed(config_manager)
-    root_state, recovered_stale_mode = _recover_stale_write_blocking_mode(config_manager, root_state)
-    root_changed = False
-    app_root = str(config_manager.app_docs_dir)
-    if root_state.get("current_root") != app_root:
-        root_state["current_root"] = app_root
-        root_changed = True
-    if not root_state.get("last_known_good_root"):
-        root_state["last_known_good_root"] = app_root
-        root_changed = True
-    if not root_state.get("last_successful_boot_at"):
-        root_state["last_successful_boot_at"] = ""
-        root_changed = True
-    if legacy_import.get("source"):
-        root_state["last_migration_source"] = str(legacy_import["source"])
-        root_state["last_migration_result"] = str(legacy_import.get("result") or "")
-        root_changed = True
-        if legacy_import.get("backup_path"):
-            root_state["last_migration_backup"] = str(legacy_import["backup_path"])
-            root_changed = True
-    elif recovered_stale_mode:
-        root_changed = True
-    elif not root_state.get("last_migration_result"):
-        root_state["last_migration_result"] = str(legacy_import.get("result") or "bootstrap_initialized")
-        root_changed = True
-    if root_changed:
-        config_manager.save_root_state(root_state)
-
-    cloud_state = config_manager.load_cloudsave_local_state()
-    cloud_changed = False
-    if not cloud_state.get("client_id"):
-        cloud_state["client_id"] = config_manager.build_default_cloudsave_local_state()["client_id"]
-        cloud_changed = True
-    next_seq = int(cloud_state.get("next_sequence_number") or 0)
-    if next_seq < 1:
-        cloud_state["next_sequence_number"] = 1
-        cloud_changed = True
-    if cloud_changed:
-        config_manager.save_cloudsave_local_state(cloud_state)
-
-    manifest = ensure_cloudsave_manifest(config_manager, preserve_existing_client_id=True)
-    return {
-        "root_state": config_manager.load_root_state(),
-        "cloudsave_local_state": config_manager.load_cloudsave_local_state(),
-        "manifest": manifest,
-        "legacy_import": legacy_import,
-    }
-
-
-def get_root_state(config_manager) -> dict[str, Any]:
-    return config_manager.load_root_state()
-
-
-def get_root_mode(config_manager) -> str:
-    state = get_root_state(config_manager)
-    return str(state.get("mode") or ROOT_MODE_NORMAL)
-
-
-def should_write_root_mode_normal_after_startup(root_state: dict[str, Any] | None) -> bool:
-    """Return True only when startup bootstrap has already settled back to normal mode."""
-    state = root_state if isinstance(root_state, dict) else {}
-    return str(state.get("mode") or ROOT_MODE_NORMAL) == ROOT_MODE_NORMAL
-
-
-def set_root_mode(config_manager, mode: str, **updates: Any) -> dict[str, Any]:
-    state = get_root_state(config_manager)
-    state["mode"] = str(mode or ROOT_MODE_NORMAL)
-    for key, value in updates.items():
-        if value is not None:
-            state[key] = value
-    config_manager.save_root_state(state)
-    return state
-
-
-def is_write_fence_active(config_manager) -> bool:
-    return get_root_mode(config_manager) in WRITE_BLOCKING_MODES
-
-
-def assert_cloudsave_writable(config_manager, *, operation: str = "write", target: str = "") -> None:
-    if is_cloudsave_disabled_due_to_local_state_unavailable():
-        return
-    mode = get_root_mode(config_manager)
-    if mode in WRITE_BLOCKING_MODES:
-        raise MaintenanceModeError(mode, operation=operation, target=target)
-
-
-def maintenance_error_payload(exc: MaintenanceModeError) -> dict[str, Any]:
-    return {
-        "success": False,
-        "error": exc.code,
-        "code": exc.code,
-        "mode": exc.mode,
-        "operation": exc.operation,
-        "target": exc.target,
-        "retryable": True,
-    }
-
-
-def scan_for_sensitive_values(payload: Any, *, path: str = "$") -> list[str]:
-    """Scan nested payloads for obviously sensitive key/value markers."""
-    findings: list[str] = []
-
-    if isinstance(payload, dict):
-        for key, value in payload.items():
-            key_str = str(key)
-            normalized_key = re.sub(r"[\s\-]+", "_", key_str.strip().lower())
-            normalized_key = re.sub(r"_+", "_", normalized_key).strip("_")
-            if normalized_key in SENSITIVE_KEY_NAMES:
-                findings.append(f"{path}.{key_str}")
-            findings.extend(scan_for_sensitive_values(value, path=f"{path}.{key_str}"))
-        return findings
-
-    if isinstance(payload, list):
-        for index, item in enumerate(payload):
-            findings.extend(scan_for_sensitive_values(item, path=f"{path}[{index}]"))
-        return findings
-
-    if isinstance(payload, str):
-        value = payload.strip()
-        if any(pattern.search(value) for pattern in SENSITIVE_VALUE_PATTERNS):
-            findings.append(path)
-    return findings
-
-
-def _cloud_apply_mutex_name(config_manager) -> str:
-    digest = hashlib.sha1(str(config_manager.app_docs_dir).encode("utf-8")).hexdigest()[:12]
-    return rf"Global\NEKO_CLOUD_APPLY_LOCK_{digest}"
-
-
-def acquire_cloud_apply_lock(config_manager) -> bool:
-    """Acquire the cross-process cloud apply lock used by maintenance mode."""
-    global _cloud_apply_lock_handle, _cloud_apply_lock_file
-
-    config_manager.ensure_local_state_directory()
-    if sys.platform == "win32":
-        try:
-            import ctypes
-
-            kernel32 = ctypes.windll.kernel32
-            ERROR_ALREADY_EXISTS = 183
-            handle = kernel32.CreateMutexW(None, True, _cloud_apply_mutex_name(config_manager))
-            last_err = kernel32.GetLastError()
-            if handle != 0:
-                if last_err != ERROR_ALREADY_EXISTS:
-                    _cloud_apply_lock_handle = handle
-                    return True
-                kernel32.CloseHandle(handle)
-                return False
-            return False
-        except Exception:
-            return True
-
-    lock_file = None
-    try:
-        import fcntl
-
-        lock_path = config_manager.local_state_dir / "cloud_apply.lock"
-        lock_file = open(lock_path, "w", encoding="utf-8")
-        try:
-            fcntl.flock(lock_file.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
-            lock_file.write(str(os.getpid()))
-            lock_file.flush()
-        except (OSError, IOError):
-            lock_file.close()
-            return False
-        _cloud_apply_lock_file = lock_file
-        return True
-    except Exception:
-        if lock_file is not None and lock_file is not _cloud_apply_lock_file:
-            try:
-                lock_file.close()
-            except Exception:
-                # Best-effort cleanup only; the acquisition fallback below keeps
-                # the existing fail-open behavior when cleanup itself fails.
-                pass
-        return True
-
-
-def release_cloud_apply_lock(config_manager) -> None:
-    global _cloud_apply_lock_handle, _cloud_apply_lock_file
-
-    if sys.platform == "win32":
-        if _cloud_apply_lock_handle is None:
-            return
-        try:
-            import ctypes
-
-            kernel32 = ctypes.windll.kernel32
-            kernel32.ReleaseMutex(_cloud_apply_lock_handle)
-            kernel32.CloseHandle(_cloud_apply_lock_handle)
-        except Exception:
-            pass
-        _cloud_apply_lock_handle = None
-        return
-
-    if _cloud_apply_lock_file is None:
-        return
-    try:
-        import fcntl
-
-        fcntl.flock(_cloud_apply_lock_file.fileno(), fcntl.LOCK_UN)
-        _cloud_apply_lock_file.close()
-    except Exception:
-        pass
-    _cloud_apply_lock_file = None
-    try:
-        os.unlink(config_manager.local_state_dir / "cloud_apply.lock")
-    except Exception:
-        pass
-
-
-def _process_holds_cloud_apply_lock() -> bool:
-    return _cloud_apply_lock_handle is not None or _cloud_apply_lock_file is not None
-
-
-def _should_preserve_write_blocking_mode(config_manager, root_state: dict[str, Any]) -> bool:
-    current_mode = str(root_state.get("mode") or ROOT_MODE_NORMAL)
-    if current_mode == ROOT_MODE_DEFERRED_INIT:
-        # 恢复态必须显式交给存储引导流程处理，不能在启动 bootstrap 里静默放行为 normal。
-        return True
-
-    if current_mode != ROOT_MODE_MAINTENANCE_READONLY:
-        return False
-
-    # 真相源是 storage_migration.json 的 pending 状态：迁移真在跑就保留 readonly，
-    # 否则视为孤儿态自愈。``last_migration_result`` 字段（含 ``restart_pending:``
-    # 前缀）只是描述上一次操作意图，不该被当作"还在进行中"的硬证据——marker
-    # 在 launcher 接力跑完迁移时才会被覆盖，任何让 launcher 跑不到那一步的事件
-    # （shutdown fire-and-forget 后 launcher 被绕过 / 半途强杀 / 迁移文件已被
-    # 善后删除）都会让 marker 残留，配合旧逻辑就把进程永久钉在 readonly 上、
-    # memory server 所有写盘静默失败。
-    try:
-        from utils.storage_migration import is_storage_migration_pending, load_storage_migration
-
-        migration_payload = load_storage_migration(config_manager)
-    except Exception as exc:
-        logger.warning("failed to load storage migration while preserving write-blocking mode: %s", exc)
-        return True
-
-    return bool(migration_payload) and is_storage_migration_pending(migration_payload)
-
-
-def _recover_stale_write_blocking_mode(config_manager, root_state: dict[str, Any]) -> tuple[dict[str, Any], bool]:
-    current_mode = str(root_state.get("mode") or ROOT_MODE_NORMAL)
-    if current_mode not in WRITE_BLOCKING_MODES:
-        return root_state, False
-
-    if _should_preserve_write_blocking_mode(config_manager, root_state):
-        return root_state, False
-
-    if _process_holds_cloud_apply_lock():
-        return root_state, False
-
-    if not acquire_cloud_apply_lock(config_manager):
-        return root_state, False
-
-    try:
-        recovered_state = dict(root_state)
-        recovered_state["mode"] = ROOT_MODE_NORMAL
-        recovered_state["last_migration_result"] = f"recovered_stale_mode:{current_mode}"
-        config_manager.save_root_state(recovered_state)
-        return recovered_state, True
-    finally:
-        release_cloud_apply_lock(config_manager)
-
-
-@contextmanager
-def cloud_apply_fence(config_manager, *, mode: str = ROOT_MODE_MAINTENANCE_READONLY, reason: str = ""):
-    """Acquire the global cloud apply lock and switch root_state into maintenance."""
-    _ensure_local_state_directory_or_raise(config_manager, "entering cloud_apply_fence")
-
-    previous_state = get_root_state(config_manager)
-    previous_mode = str(previous_state.get("mode") or ROOT_MODE_NORMAL)
-    if not acquire_cloud_apply_lock(config_manager):
-        raise MaintenanceModeError(
-            get_root_mode(config_manager),
-            operation="acquire_lock",
-            target="cloud_apply_lock",
-        )
-    try:
-        set_root_mode(
-            config_manager,
-            mode,
-            last_migration_result=reason or previous_state.get("last_migration_result", ""),
-        )
-        yield get_root_state(config_manager)
-    finally:
-        try:
-            set_root_mode(config_manager, previous_mode)
-        finally:
-            release_cloud_apply_lock(config_manager)
